% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_COOBS.R
\name{analyze_COOBS}
\alias{analyze_COOBS}
\title{Perform the COunt of ObServations (COOBS) algorithm using existing site data
and raster metrics}
\usage{
analyze_COOBS(
  mraster = NULL,
  existing = NULL,
  cores = 1,
  threshold = 0.95,
  plot = FALSE
)
}
\arguments{
\item{mraster}{spatRaster. ALS metrics raster.}

\item{existing}{sf or data.frame.  Existing plot network.}

\item{threshold}{Numeric. Proxy maximum pixel quantile to avoid outliers. \code{default = 0.95}}

\item{plot}{Logical. Plots output strata raster and visualized
strata with boundary dividers.}
}
\value{

}
\description{
Perform the COunt of ObServations (COOBS) algorithm using existing site data
and raster metrics
}
\seealso{
Other analyze functions: 
\code{\link{analyze_popLHC}()},
\code{\link{analyze_sampOptLHC}()}
}
\concept{analyze functions}
