% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_HELS.R
\name{analyze_HELS}
\alias{analyze_HELS}
\title{Perform the Hypercube Evaluation of a Legacy Sample (HELS) algorithm using existing site data
and raster metrics. New samples are allocated based on quantile ratios between the existing sample
and covariate dataset.}
\usage{
analyze_HELS(
  mraster = NULL,
  existing = NULL,
  nQuant = 20,
  nSamp = 100,
  plot = FALSE
)
}
\arguments{
\item{mraster}{spatRaster. ALS metrics raster.}

\item{existing}{sf or data.frame.  Existing plot network.}

\item{nSamp}{Numeric. Maximum number of new samples to allocate.}

\item{plot}{Logical. Plots output strata raster and visualized
strata with boundary dividers.}
}
\value{

}
\description{
Perform the Hypercube Evaluation of a Legacy Sample (HELS) algorithm using existing site data
and raster metrics. New samples are allocated based on quantile ratios between the existing sample
and covariate dataset.
}
\seealso{
Other analyze functions: 
\code{\link{analyze_COOBS}()},
\code{\link{analyze_HELSrem}()},
\code{\link{analyze_popLHC}()},
\code{\link{analyze_sampOptLHC}()}
}
\concept{analyze functions}
