% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_HELSrem.R
\name{analyze_HELSrem}
\alias{analyze_HELSrem}
\title{Evaluates existing sample data using an adapted Hypercube Evaluation of a Legacy Sample (HELS) algorithm
to remove redundant samples based on covariate and sample quantile comparisons}
\usage{
analyze_HELSrem(mraster = mraster, existing = existing, nQuant = nQuant)
}
\arguments{
\item{mraster}{spatRaster. ALS metrics raster.}

\item{existing}{sf or data.frame.  Existing plot network.}

\item{threshold}{Numeric. Proxy maximum pixel quantile to avoid outliers. \code{default = 0.95}}
}
\value{

}
\description{
Evaluates existing sample data using an adapted Hypercube Evaluation of a Legacy Sample (HELS) algorithm
to remove redundant samples based on covariate and sample quantile comparisons
}
\seealso{
Other analyze functions: 
\code{\link{analyze_popLHC}()},
\code{\link{analyze_sampOptLHC}()}
}
\concept{analyze functions}
