% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_popLHC.R
\name{analyze_popLHC}
\alias{analyze_popLHC}
\title{Analyze covariates for LHC}
\usage{
analyze_popLHC(mraster, PCA = TRUE, quant = TRUE, nQuant = 20, KLdiv = TRUE)
}
\arguments{
\item{mraster}{spatRaster. ALS metrics raster.}

\item{PCA}{Logical. Calculates principal component loadings of the population for PCA similarity factor testing. 
\code{default = TRUE}.}

\item{quant}{Logical. Calculates quantile matrix of the population for quantile comparison testing. 
\code{default = TRUE}.}

\item{nQuant}{Numeric. Number of quantiles to divide the population into for \code{quant}. 
\code{default = 20}.}

\item{KLdiv}{Logical. Calculates covariate matrix of the population for Kullback–Leibler divergence testing. 
\code{default = TRUE}. Relies on \code{quant = TRUE} to calculate.}
}
\value{
List of matrices to be used as input for \code{analyze_sampOptLHC}.
}
\description{
Population level analysis of metric raster data to determine optimal Latin Hypercube sample size
}
\seealso{
Other analyze functions: 
\code{\link{analyze_HELSrem}()},
\code{\link{analyze_sampOptLHC}()}
}
\concept{analyze functions}
