% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_sampOptLHC.R
\name{analyze_sampOptLHC}
\alias{analyze_sampOptLHC}
\title{Population level analysis of metric raster data to determine optimal Latin Hypercube sample size}
\usage{
analyze_sampOptLHC(
  popLHC = NULL,
  PCA = TRUE,
  quant = TRUE,
  KLdiv = TRUE,
  minSamp = 10,
  maxSamp = 100,
  step = 10,
  rep = 10,
  iter = 10000
)
}
\arguments{
\item{popLHC}{List. Output from \code{analyze_popLHC} function.}

\item{PCA}{Logical. Calculates principal component loadings of the population for PCA similarity factor testing. 
\code{default = TRUE}.}

\item{quant}{Logical. Calculates quantile matrix of the population for quantile comparison testing. 
\code{default = TRUE}.}

\item{KLdiv}{Logical. Calculates covariate matrix of the population for Kullback–Leibler divergence testing. 
\code{default = TRUE}. Relies on \code{quant = TRUE} to calculate.}

\item{minSamp}{Numeric. Minimum sample size to test. \code{default = 10}.}

\item{maxSamp}{Numeric. Maximum sample size to test. \code{default = 100}.}

\item{step}{Numeric. Sample step size for each iteration. \code{default = 10}.}

\item{rep}{Numeric. Internal repetitions for each sample size. \code{default = 10}.}

\item{iter}{Numeric. Internal to \code{clhs} - A positive number, giving the number of 
iterations for the Metropolis-Hastings annealing process. Defaults to \code{10000}.}
}
\value{
data.frame with summary statistics.
}
\description{
Population level analysis of metric raster data to determine optimal Latin Hypercube sample size
}
\seealso{
Other analyze functions: 
\code{\link{analyze_COOBS}()},
\code{\link{analyze_popLHC}()}
}
\concept{analyze functions}
