% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_allocation.R
\name{calculate_allocation}
\alias{calculate_allocation}
\title{Determine required samples in strata}
\usage{
calculate_allocation(
  sraster,
  nSamp,
  allocation = "prop",
  mraster = NULL,
  metric = NULL,
  existing = NULL,
  force = FALSE
)
}
\arguments{
\item{sraster}{spatRaster. Stratification raster to be used for sampling.}

\item{nSamp}{Numeric. Number of desired samples.}

\item{allocation}{Character. Allocation algorithm to be used. Either \code{prop} (default) for proportional allocation
or \code{optim} for optimal allocation.}

\item{mraster}{spatRaster. ALS metrics raster. Required when \code{allocation = optim}.}

\item{metric}{Numeric/Character. Index or name of primary covariate within mraster to stratify.
Required when \code{allocation = optim}.}

\item{existing}{sf or data.frame.  Existing plot network.}

\item{force}{Logical. \code{Default = FALSE} - force \code{nSamp} to be exactly the user defined value
in cases where nSamp and \code{sraster} strata count are not equally divisible. Has no effect when \code{existing}
is provided.}
}
\value{
data.frame of:
\itemize{
\item{strata} - Strata ID.
\item{total} - Total samples to be allocated (positive) or removed (negative).
\item{need} - Total required samples per strata.
}
}
\description{
Determine required samples in strata
}
\details{
Determine how many samples to allocate within strata
}
\section{Details}{

\itemize{
\item{nSamp = Desired sample count}
\item{Nh = Count of cells within strata 'h'}
\item{N = Total cound of cells}
\item{\eqn{\sigma}h = Sample standard deviation}
\item{nh = samples allocated with strata 'h'}
}

\eqn{nh = nSamp * Nh / N}

\eqn{nh = nSamp  Nh * Nh * \sigma h / \sum Nk * \sigma k}
}

\examples{
#--- Load strata raster and existing samples---#
r <- system.file("extdata", "kmeans.tif", package = "sgsR")
sr <- terra::rast(r)

e <- system.file("extdata", "existing.shp", package = "sgsR")
e <- sf::st_read(e)

#--- perform grid sampling ---#
calculate_allocation(sraster = sr, 
                     nSamp = 200)

e.sr <- extract_strata(sraster = sr, 
                       existing = e)

calculate_allocation(sraster = sr, 
                     nSamp = 200, 
                     existing = e.sr)
                     
}
\seealso{
Other calculate functions: 
\code{\link{calculate_coobs}()},
\code{\link{calculate_distance}()},
\code{\link{calculate_lhsPop}()},
\code{\link{calculate_pcomp}()}
}
\author{
Tristan R.H. Goodbody
}
\concept{calculate functions}
