% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance.R
\name{calculate_distance}
\alias{calculate_distance}
\title{Distance to access layer}
\usage{
calculate_distance(
  raster,
  access,
  plot = FALSE,
  filename = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{raster}{spatRaster. Raster to be used to calculate pixel level distance to access layer.}

\item{access}{sf. Road access network - must be lines.}

\item{plot}{Logical. Plots output strata raster with samples.}

\item{filename}{Character. Path to write output samples.}

\item{overwrite}{Logical. Choice to overwrite existing \code{filename} if it exists.}
}
\value{
Input raster with \code{dist2access} layer appended.
}
\description{
Per pixel distance to nearest access vector. Intended to be used as a `cost` constraint
within the \code{\link{sample_clhs}} function
}
\examples{
#--- Load raster and access files ---#
r <- system.file("extdata", "kmeans.tif", package = "sgsR")
sr <- terra::rast(r)

a <- system.file("extdata", "roads.shp", package = "sgsR")
ac <- sf::st_read(a)

calculate_distance(
  raster = sr,
  access = ac,
  plot = TRUE
)

calculate_distance(
  raster = sr,
  access = ac,
  plot = TRUE,
  filename = tempfile(fileext = ".tif")
)
}
\seealso{
Other calculate functions: 
\code{\link{calculate_allocation}()},
\code{\link{calculate_coobs}()},
\code{\link{calculate_lhsPop}()},
\code{\link{calculate_pcomp}()}
}
\concept{calculate functions}
