% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lhsOpt.R
\name{calculate_lhsOpt}
\alias{calculate_lhsOpt}
\title{Analyze optimal Latin hypercube sample number}
\usage{
calculate_lhsOpt(
  popLHS = NULL,
  PCA = TRUE,
  quant = TRUE,
  KLdiv = TRUE,
  minSamp = 10,
  maxSamp = 100,
  step = 10,
  rep = 10,
  iter = 10000
)
}
\arguments{
\item{popLHS}{List. Output from \code{\link{calculate_lhsPop}} function.}

\item{PCA}{Logical. Calculates principal component loadings of the population for PCA similarity factor testing.
\code{default = TRUE}.}

\item{quant}{Logical. Calculates quantile matrix of the population for quantile comparison testing.
\code{default = TRUE}.}

\item{KLdiv}{Logical. Calculates covariate matrix of the population for Kullback–Leibler divergence testing.
\code{default = TRUE}. Relies on \code{quant = TRUE} to calculate.}

\item{minSamp}{Numeric. Minimum sample size to test. \code{default = 10}.}

\item{maxSamp}{Numeric. Maximum sample size to test. \code{default = 100}.}

\item{step}{Numeric. Sample step size for each iteration. \code{default = 10}.}

\item{rep}{Numeric. Internal repetitions for each sample size. \code{default = 10}.}

\item{iter}{Numeric. Internal to \code{\link[clhs]{clhs}} - A positive number, giving the number of
iterations for the Metropolis-Hastings annealing process. Defaults to \code{10000}.}
}
\value{
data.frame with summary statistics.
}
\description{
Population level analysis of metric raster data to determine optimal Latin Hypercube sample size
}
\examples{
#--- Load raster and access files ---#
r <- system.file("extdata","wall_metrics_small.tif", package = "sgsR")
mr <- terra::rast(r)

#--- calculate lhsPop details ---#
poplhs <- calculate_lhsPop(mraster = mr)

calculate lhsOpt(popLHS = poplhs)

calculate lhsOpt(popLHS = poplhs, PCA = FALSE, iter = 200)

}
\references{
Malone BP, Minansy B, Brungard C. 2019. Some methods to improve the utility of conditioned Latin hypercube sampling. PeerJ 7:e6451 DOI 10.7717/peerj.6451
}
\concept{analyze functions}
