% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pcomp.R
\name{calculate_pcomp}
\alias{calculate_pcomp}
\title{Raster principal components}
\usage{
calculate_pcomp(
  mraster = NULL,
  nComp = NULL,
  center = TRUE,
  scale = TRUE,
  plot = FALSE,
  details = FALSE,
  ...
)
}
\arguments{
\item{mraster}{spatRaster. ALS metrics raster.}

\item{nComp}{Numeric. Value indicating number of principal components to be rasterized.
prior to analysis.}

\item{center}{Logical. Value indicating whether the variables should be shifted to be zero centered.}

\item{scale}{Logical. Value indicating whether the variables should be scaled to have unit variance}

\item{plot}{Logical. Plots output strata raster with samples.}

\item{details}{Logical. If \code{FALSE} (default) output is only
 stratification raster. If \code{TRUE} return a list
where \code{$details} is additional stratification information and
 \code{$raster} is the output stratification spatRaster.
 @param ... Additional arguments to be passed to \code{\link[stats]{kmeans}} function.}

\item{...}{Additional arguments to be passed to \code{\link[stats]{prcomp}}.}
}
\value{
Output raster with specified number of principal components as layers.
}
\description{
Calculate and rasterize principal components from a metric raster
}
\examples{
#--- Load raster ---#
r <- system.file("extdata","wall_metrics_small.tif", package = "sgsR")
mr <- terra::rast(r)

calculate_pcomp(mraster = mr, nComp = 5)

calculate_pcomp(mraster = mr, nComp = 3, plot = TRUE, details = TRUE)

}
\seealso{
Other calculate functions: 
\code{\link{calculate_COOBS}()},
\code{\link{calculate_distance}()},
\code{\link{calculate_lhsPop}()},
\code{\link{calculate_reqSamples}()}
}
\concept{calculate functions}
