% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_representation.R
\name{calculate_representation}
\alias{calculate_representation}
\title{Compare representation of samples within sraster strata}
\usage{
calculate_representation(sraster, existing, plot = FALSE)
}
\arguments{
\item{sraster}{spatRaster. Stratification raster to be used for sampling.}

\item{existing}{sf or data.frame.  Existing plot network.}

\item{plot}{Logical. Plot histogram output displaying the frequency of strata coverage and sampling coverage for \code{sraster} and \code{existing}.}
}
\value{
Returns a tibble where:
\itemize{
\item{strata} - \code{sraster} strata ID.
\item{srasterFreq} - Coverage frequency percent of \code{sraster} strata.
\item{sampleFreq} - Sampling frequency percent within \code{sraster} strata.
\item{diffFreq} - Difference between \code{srasterFreq} & \code{sampleFreq}. Positive values indicate over representation
\item{nSamp} - Number of samples within each strata in \code{existing}.
\item{need} - \code{srasterFreq * sum(nSamp)}. Total theoretical number of required samples to be representative of strata coverage.
This values is rounded. It is important for the user to consider \code{diffFreq}. A small difference - e.g. 1 percent -
in sample vs. sraster frequency could correspond to the algorithm allocating or removing samples that could likely be ignored.
}
}
\description{
Compare representation of samples within sraster strata
}
\details{
Calculate how well sraster strata are represented in existing samples
}
\examples{
### --- generate example stratification ---###

#--- load ALS metrics from sgsR internal data ---#
r <- system.file("extdata", "wall_metrics.tif", package = "sgsR")

#--- read ALS metrics using the terra package ---#
mraster <- terra::rast(r)

#--- perform stratification ---#
sraster <- strat_kmeans(
  mraster = mraster$zq90,
  nStrata = 6,
  plot = TRUE
)

### --- create existing sample network ---###

#--- simple random sampling ---#
existing <- sample_srs(
  raster = mraster$zq90,
  nSamp = 100
)

#--- calculate representation ---#

calculate_representation(
  sraster = sraster,
  existing = existing,
  plot = TRUE
)
}
\seealso{
Other calculate functions: 
\code{\link{calculate_allocation}()},
\code{\link{calculate_coobs}()},
\code{\link{calculate_distance}()},
\code{\link{calculate_pcomp}()},
\code{\link{calculate_pop}()},
\code{\link{calculate_sampsize}()}
}
\author{
Tristan R.H. Goodbody, Martin Queinnec
}
\concept{calculate functions}
