% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_reqSamples.R
\name{calculate_reqSamples}
\alias{calculate_reqSamples}
\title{Determine required samples in strata}
\usage{
calculate_reqSamples(sraster, nSamp, existing = NULL, force = FALSE)
}
\arguments{
\item{sraster}{spatRaster. Stratification raster to be used for sampling.}

\item{nSamp}{Numeric. Number of desired samples.}

\item{existing}{sf or data.frame.  Existing plot network.}

\item{force}{Logical. \code{Default = FALSE} - force \code{nSamp} to be exactly the user defined value
in cases where nSamp and \code{sraster} strata count are not equally divisible. Has no effect when \code{existing}
is provided.}
}
\value{
data.frame of:
\itemize{
\item{strata} - Strata ID.
\item{total} - Total samples to be allocated (positive) or removed (negative).
\item{need} - Total required samples per strata.
}
}
\description{
Determine required samples in strata
}
\details{
Count the required number of samples for each raster strata
}
\examples{
#--- Load strata raster and existing samples---#
r <- system.file("extdata", "kmeans.tif", package = "sgsR")
sr <- terra::rast(r)

e <- system.file("extdata", "existing.shp", package = "sgsR")
e <- sf::st_read(e)

#--- perform grid sampling ---#
calculate_reqSamples(sraster = sr, 
                     nSamp = 200)

e.sr <- extract_strata(sraster = sr, 
                       existing = e)

calculate_reqSamples(sraster = sr, 
                     nSamp = 200, 
                     existing = e.sr)
                     
}
\seealso{
Other calculate functions: 
\code{\link{calculate_coobs}()},
\code{\link{calculate_distance}()},
\code{\link{calculate_lhsPop}()},
\code{\link{calculate_pcomp}()}
}
\author{
Tristan R.H. Goodbody
}
\concept{calculate functions}
