% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-existing.R
\name{coords_existing}
\alias{coords_existing}
\title{Get existing and XY coordinates}
\usage{
coords_existing(existing)
}
\arguments{
\item{existing}{sf 'POINT'.  Existing plot network.}
}
\value{
\code{existing} with respective 'X' and 'Y' columns representing the coordinates of the input \code{sf} object.
}
\description{
This function extracts the 'X' and 'Y' coordinates from an 'sf' object and returns them in a data.frame format.
}
\keyword{internal}
