% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-extract.R
\name{extract}
\alias{extract}
\alias{extract_strata}
\alias{extract_metrics}
\title{Extract}
\usage{
extract_strata(
  sraster,
  existing,
  data.frame = FALSE,
  filename = NULL,
  overwrite = FALSE
)

extract_metrics(
  mraster,
  existing,
  data.frame = FALSE,
  filename = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{sraster}{spatRaster. Stratification raster.}

\item{existing}{sf. Samples resulting from sample_* functions.}

\item{data.frame}{Logical. If true outputs as data.frame}

\item{filename}{Character. Path to write stratified raster to disc.}

\item{overwrite}{Logical. Specify whether \code{filename} should be overwritten on disc.}

\item{mraster}{Spatraster. Primary covariate raster to stratify.}
}
\value{
An sf or data.frame object of samples with strata attributes

An sf or data.frame object of samples with associated raster cell attributes
}
\description{
Extract raster values to samples

Extract metric raster attributes to existing
}
\concept{extract}
\keyword{internal}
