% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_metrics.R
\name{extract_metrics}
\alias{extract_metrics}
\title{Extract raster cell values to samples}
\usage{
extract_metrics(raster, samples, data.frame = FALSE)
}
\arguments{
\item{raster}{spatRaster. Raster to be used for simple random sampling}

\item{samples}{sf. Samples resulting from sample_* functions.}

\item{data.frame}{Logical. If true outputs as data.frame}
}
\value{
An sf or data.frame object with raster cell attributes
}
\description{
Extract raster cell values to samples
}
\examples{
extract_metrics(raster = raster, samples = samples)
extract_metrics(raster = raster, samples = samples, data.frame = TRUE)

}
\seealso{
Other extract functions: 
\code{\link{extract_existing}()}
}
\concept{extract functions}
