% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mask.R
\name{masking}
\alias{masking}
\alias{mask_access}
\alias{mask_existing}
\title{Masking}
\usage{
mask_access(raster, access, buff_inner = NULL, buff_outer)

mask_existing(access, existing, buff_inner = NULL, buff_outer)
}
\arguments{
\item{raster}{SpatRaster. Raster to be masked.}

\item{access}{sf 'LINESTRING' or 'MULTILINESTRING'. Access network.}

\item{buff_inner}{Numeric. Inner buffer boundary specifying distance
from access where plots cannot be sampled.}

\item{buff_outer}{Numeric. Outer buffer boundary specifying distance
from access where plots can be sampled.}

\item{existing}{sf 'POINT'.  Existing plot network.}
}
\value{
Raster/existing samples masked by provided \code{access} layer and buffers.
}
\description{
Mask covariates and existing samples by access
}
\concept{masking}
\keyword{internal}
