% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mat.R
\name{matrices}
\alias{matrices}
\alias{mat_quant}
\alias{mat_cov}
\alias{mat_covNB}
\title{Matrices}
\usage{
mat_quant(vals, nQuant, nb)

mat_cov(vals, nQuant, nb, matQ)

mat_covNB(vals, nQuant, nb, matQ)
}
\arguments{
\item{vals}{Covariate / sample data}

\item{nQuant}{Number of quantiles to be used}

\item{nb}{Number of bands}

\item{matQ}{Quantile matrix}
}
\description{
Create covariate and sample matrices
}
\concept{matrices}
\keyword{#---}
\keyword{---#}
\keyword{a}
\keyword{bar---#}
\keyword{calculating}
\keyword{covariance}
\keyword{for}
\keyword{function}
\keyword{internal}
\keyword{matrix}
\keyword{progress}
\keyword{quantile}
\keyword{with}
\keyword{without}
