% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{plot}
\alias{plot}
\alias{classPlot}
\title{Plot}
\usage{
classPlot(dfc, coordsgrps, metric, metric2, samp)
}
\arguments{
\item{dfc}{data.frame. Values for metric1 and metric2}

\item{coordsgrps}{List. Cartesian coordinates of each strata}

\item{metric}{Numeric/Character. Index or name of primary covariate within \code{mraster} to stratify.}

\item{metric2}{Numeric/Character. Index or name of secondary covariate within \code{mraster} to stratify.}

\item{samp}{Numeric. Determines proportion of cells to plot}
}
\description{
Plot

Class Plot
}
\concept{plot}
\keyword{internal}
