% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataio.R
\name{readGenepop}
\alias{readGenepop}
\title{Read Genepop input from file and convert to  sgsObj object}
\usage{
readGenepop(path_to_genefile, missing_vals = "-999")
}
\arguments{
\item{path_to_genefile}{A file path pointing to Genepop input file. Must point to a text file in following format in Genepop manual (see details below).}

\item{missing_vals}{Value indidicating missing data. Default is -999}
}
\value{
An object of the class sgsObj, which is used for futher sgs analyses.
}
\description{
This function reads a Genepop formatted text file and converts it to an sgsObj.
}
\section{Details}{


Genepop is a popular program for calculating various tests on genetic data \href{http://genepop.curtin.edu.au/}{here}.

Data format follows that found at \href{http://genepop.curtin.edu.au/help_input.html#Input}{here}.

Generally, the first line represents information about data or data title.

The second line begins the names of the first locus. Each following locus is given its own line.

An alternative formatting separates loci by commas, however this function does not work with this style.

Each line of data MUST contain an identification, which is not required for general Genepop files.

Each individual is given its own line, with loci data following identification.

Different populations are separated by some form of POP.

\emph{Note:} This function as been tested only for diploid genetic data.
}
\examples{
## Read in example data provided with package
path_to_example_data <- system.file("extdata", "example_genepop.txt", package = "sgsR")

## Use readGenepop to read and convert data to sgsObj
dat <- readGenepop(path_to_example_data, missing_vals = "0")

dat
}

