% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataio.R
\name{readSpagedi}
\alias{readSpagedi}
\title{Read SPAGeDi input from file and convert to sgsObj object}
\usage{
readSpagedi(path_to_spagedi_file, missing_val = "-999")
}
\arguments{
\item{path_to_spagedi_file}{A file path pointing to SPAGeDi input file. Must point to a tab delimited text file in following format in SPAGeDi manual (see details below).}

\item{missing_val}{Value indidicating missing data. Default is -999}
}
\value{
An object of the class sgsObj, which is used for futher sgs analyses.
}
\description{
This function reads a SPAGeDi formatted text file (tab delimited text) and converts it to an sgsObj.
}
\section{Details}{


SPAGeDi (Spatial Pattern Analysis of Genetic Diversity) is a popular program for calculating spatial genetic structure, among other things. More information about SPAGeDi can be found at their hompage \href{http://ebe.ulb.ac.be/ebe/SPAGeDi.html}{here}.

Data format should follow the style described in section 3.1 of the SPAGeDi manual.

File should be a tab delimited text file.

Briefly, the first line is a series of 6 numbers separated by tabs: number of individuals, number of categories, number of spatial coordinates, number of loci, number of digits used to code one allele, and ploidy.

The second line indicates the distance intervals (which can be changed after inputting file).

The third line lists the column names for individuals, categories, spatial coordinates, and loci.

The fourth line begins the sample ID, category, spatial coordinate, and genotype data, with each line representing an individual.

END is entered in the line after the last individual.

\emph{Note:} This function as been tested only for diploid genetic data in 2d coordinate systems.
}
\examples{
\dontrun{
## Read in example data provided with package
path_to_example_data <- system.file("extdata", "obataua_spagedi.txt", package = "sgsR")

## Use readSpagedi to read and convert data to sgsObj
dat <- readSpagedi(path_to_example_data, missing_val = "0")

dat
}
}

