% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_clhs.R
\name{sample_clhs}
\alias{sample_clhs}
\title{CLHS sampling}
\usage{
sample_clhs(
  mraster,
  nSamp,
  existing = NULL,
  access = NULL,
  buff_inner = NULL,
  buff_outer = NULL,
  plot = FALSE,
  details = FALSE,
  ...
)
}
\arguments{
\item{mraster}{spatRaster. ALS metrics raster.}

\item{nSamp}{Numeric. Number of desired samples.}

\item{existing}{sf. Samples resulting from sample_* functions.}

\item{access}{sf. Road access network - must be lines.}

\item{buff_inner}{Numeric. Inner buffer boundary specifying distance
from access where plots cannot be sampled.}

\item{buff_outer}{Numeric. Outer buffer boundary specifying distance
from access where plots can be sampled.}

\item{plot}{Logical. Plots output strata raster with samples.}

\item{details}{Logical. If \code{FALSE} (default) output is only
 stratification raster. If \code{TRUE} return a list
where \code{$details} is additional stratification information and
 \code{$raster} is the output stratification spatRaster.
 @param ... Additional arguments to be passed to \code{\link[stats]{kmeans}} function.}

\item{...}{Additional arguments for clhs sampling. See \code{\link[clhs]{clhs}}.}
}
\value{
An sf object with \code{nSamp} stratified samples.
}
\description{
Conditioned Latin Hypercube Sampling using the \code{\link[clhs]{clhs-package}} package functionality
}
\seealso{
Other sample functions: 
\code{\link{sample_aHELS}()},
\code{\link{sample_balanced}()},
\code{\link{sample_grid}()},
\code{\link{sample_srs}()},
\code{\link{sample_strat}()}
}
\concept{sample functions}
