% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_existing_srs.R
\name{sample_existing_srs}
\alias{sample_existing_srs}
\title{Randomly sample from an existing dataset}
\usage{
sample_existing_srs(existing, nSamp, filename = NULL, overwrite = FALSE)
}
\arguments{
\item{existing}{sf 'POINT'.  Existing plot network.}

\item{nSamp}{Numeric. Number of desired samples.}

\item{filename}{Character. Path to write output samples.}

\item{overwrite}{Logical. Choice to overwrite existing \code{filename} if it exists.}
}
\value{
A data frame or spatial data frame containing the random sample.
}
\description{
This function takes an existing \code{sf} object and returns a random sub-sample of size \code{nSamp}.
}
\keyword{internal}
