% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_existing_strat.R
\name{sample_existing_strat}
\alias{sample_existing_strat}
\title{Sample Existing Data Based on Strata}
\usage{
sample_existing_strat(
  existing,
  toSample,
  nSamp,
  filename = NULL,
  overwrite = NULL
)
}
\arguments{
\item{existing}{A data frame containing existing data.}

\item{toSample}{A data frame specifying the desired sample sizes for each strata.}

\item{nSamp}{An integer specifying the number of samples to take.}

\item{filename}{A character string specifying the name of the file to write the samples to,
or NULL if no file should be written.}

\item{overwrite}{A logical value indicating whether to overwrite the file if it already exists,
or NULL if the file should not be overwritten.}
}
\value{
A data frame containing a sample of the existing data based on the desired sample sizes
for each strata.
}
\description{
This function takes a data frame of existing data, a data frame of desired sample sizes
for each strata, the number of samples to take, and optionally a file name and overwrite
parameter. It returns a sample of the existing data based on the desired sample sizes for
each strata, and optionally writes the resulting samples to a file.
}
\keyword{internal}
