% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_grid.R
\name{sample_grid}
\alias{sample_grid}
\title{Grid sampling}
\usage{
sample_grid(
  raster,
  gridsize,
  access = NULL,
  buff_inner = NULL,
  buff_outer = NULL,
  plot = FALSE,
  filename = NULL,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{raster}{spatRaster. Raster used to define extent of fishnet grid}

\item{gridsize}{Numeric. Desired distance between samples}

\item{access}{sf. Road access network - must be lines.}

\item{buff_inner}{Numeric. Inner buffer boundary specifying distance
from access where plots cannot be sampled.}

\item{buff_outer}{Numeric. Outer buffer boundary specifying distance
from access where plots can be sampled.}

\item{plot}{Logical. Plots output strata raster with samples.}

\item{filename}{Character. Path to write output samples.}

\item{overwrite}{Logical. Choice to overwrite existing \code{filename} if it exists.}

\item{...}{Additional arguments for \code{\link[sf]{st_make_grid}}.}
}
\value{
An sf object with sampled points at intersections of fishnet grid.
}
\description{
Sample landscape using a fishnet grid pattern.
}
\examples{
#--- Load raster and access files ---#
r <- system.file("extdata","kmeans.tif", package = "sgsR")
sr <- terra::rast(r)

a <- system.file("extdata","roads.shp", package = "sgsR")
ac <- sf::st_read(a)

e <- system.file("extdata","existing.shp", package = "sgsR")
e <- sf::st_read(e)

#--- perform grid sampling ---#
sample_grid(raster = sr, gridsize = 200, plot = TRUE)

sample_grid(raster = sr, gridsize = 100, access = ac, 
buff_inner = 50, buff_outer = 200)

sample_grid(raster = sr, gridsize = 200, access = ac, buff_inner = 100,
 buff_outer = 400, filename = tempfile(fileext = ".shp"), plot = TRUE)

}
\seealso{
Other sample functions: 
\code{\link{sample_aHELS}()},
\code{\link{sample_balanced}()},
\code{\link{sample_clhs}()},
\code{\link{sample_srs}()},
\code{\link{sample_strat}()}
}
\author{
Tristan R.H. Goodbody
}
\concept{sample functions}
