% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_srs.R
\name{sample_srs}
\alias{sample_srs}
\title{Randomly sample within a raster extent}
\usage{
sample_srs(
  raster,
  ns,
  mindist,
  access = NULL,
  buff_inner = NULL,
  buff_outer = NULL
)
}
\arguments{
\item{raster}{spatRaster. Raster to be used for simple random sampling}

\item{ns}{Numeric. Number of desired samples.}

\item{mindist}{Numeric. Minimum allowable distance between selected samples.}

\item{access}{sf. Road access network - must be lines.}

\item{buff_inner}{Numeric. Inner buffer boundary specifying distance from access where plots cannot be sampled.}

\item{buff_outer}{Numeric. Outer buffer boundary specifying distance from access where plots can be sampled.}
}
\value{
An sf object with 'ns' randomly sampled points
}
\description{
Randomly sample within a raster extent
}
\examples{
sample_srs(raster = raster, ns = 200, mindist = 200)
sample_srs(raster = raster, ns = 200, mindist = 200, access = roads, buff_inner = 50, buff_outer = 200)

}
\seealso{
Other sample functions: 
\code{\link{sample_strat}()}
}
\concept{sample functions}
