% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_strat.R
\name{sample_strat}
\alias{sample_strat}
\title{Stratified sampling}
\usage{
sample_strat(
  sraster,
  nSamp,
  mindist = NULL,
  existing = NULL,
  include = FALSE,
  access = NULL,
  buff_inner = NULL,
  buff_outer = NULL,
  force = FALSE,
  wrow = 3,
  wcol = 3,
  plot = FALSE,
  details = FALSE,
  filename = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{sraster}{spatRaster. Stratification raster to be used for sampling.}

\item{nSamp}{Numeric. Number of desired samples. \code{existing include} and \code{force} influence this value.}

\item{mindist}{Numeric. Minimum allowable distance between selected
samples. Default = 100.}

\item{existing}{sf or data.frame.  Existing plot network.}

\item{include}{Logical. If \code{TRUE} include existing plots in \code{nSamp} total.}

\item{access}{sf. Road access network - must be lines.}

\item{buff_inner}{Numeric. Inner buffer boundary specifying distance
from access where plots cannot be sampled.}

\item{buff_outer}{Numeric. Outer buffer boundary specifying distance
from access where plots can be sampled.}

\item{force}{Logical. \code{Default = FALSE} - force \code{nSamp} to be exactly the user defined value
in cases where nSamp and \code{sraster} strata count are not equally divisible. Has no effect when \code{existing}
is provided.}

\item{wrow}{Numeric. Number of row in the focal window (default is 3).}

\item{wcol}{Numeric. Number of columns in the focal window (default is 3).}

\item{plot}{Logial. Plots existing (circles) and new (crosses) samples.}

\item{details}{Logical. If \code{FALSE} (default) output is sf object of
stratified samples. If \code{TRUE} return a list
where \code{$details} additional sampling information and \code{$raster}
is an sf object of stratified samples.}

\item{filename}{Character. Path to write output samples.}

\item{overwrite}{Logical. Choice to overwrite existing \code{filename} if it exists.}
}
\value{
An sf object with \code{nSamp} stratified samples.
}
\description{
Sampling based on a stratified raster.
}
\note{
The sampling is performed in 2 stages:
\enumerate{

\item \code{Rule 1} - Sample within grouped stratum pixels defined within the
\code{wrow, wcol} parameters

\item \code{Rule 2} - If no more samples exist to satisfy desired sampling count,
 individual stratum pixels are sampled.
 
 The rule applied to a allocate each sample is defined in the \code{rule} attribute of output samples.

}
}
\examples{
#--- Load raster and access files ---#
r <- system.file("extdata", "kmeans.tif", package = "sgsR")
sr <- terra::rast(r)

a <- system.file("extdata", "roads.shp", package = "sgsR")
ac <- sf::st_read(a)

e <- system.file("extdata", "existing.shp", package = "sgsR")
e <- sf::st_read(e)

sample_strat(sraster = sr, 
             nSamp = 200, 
             plot = TRUE)
             
#--- extract strata values to existing samples ---#              
e.sr <- extract_strata(sraster = sr, existing = e)              

sample_strat(sraster = sr, 
             nSamp = 200, 
             access = ac,
             existing = e.sr, 
             mindist = 200, 
             buff_inner = 50, 
             buff_outer = 200)

sample_strat(sraster = sr, 
             nSamp = 200, 
             access = ac,
             buff_inner = 50, 
             buff_outer = 200, 
             filename = tempfile(fileext = ".shp"))
             
}
\seealso{
Other sample functions: 
\code{\link{sample_ahels}()},
\code{\link{sample_balanced}()},
\code{\link{sample_clhs}()},
\code{\link{sample_grid}()},
\code{\link{sample_srs}()}
}
\author{
Tristan R.H. Goodbody & Martin Queinnec
}
\concept{sample functions}
