% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genetic_functions.R
\name{sgs}
\alias{sgs}
\title{Title}
\usage{
sgs(sgsObj, distance_intervals, nperm = 999)
}
\arguments{
\item{sgsObj}{}

\item{distance_intervals}{}

\item{nperm}{}
}
\value{
1
}
\description{
Title
}
\examples{
1
}

