% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgsR-package.R
\docType{package}
\name{sgsR-package}
\alias{sgsR}
\alias{sgsR-package}
\title{sgsR: sgsR: Structurally Guided Sampling Approaches using ALS Data for R}
\description{
This package provides functions to enable stratification and sampling for structurally guided sampling using ALS data.
}
\keyword{internal}
