% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_fri.R
\name{strat_fri}
\alias{strat_fri}
\title{Stratify Forest Resources Inventory}
\usage{
strat_fri(
  FRI,
  attribute,
  features,
  raster,
  map = FALSE,
  stack = FALSE,
  filename = NULL,
  overwrite = FALSE,
  plot = FALSE,
  details = FALSE
)
}
\arguments{
\item{FRI}{sf. Forest resources inventory polygon coverage.}

\item{attribute}{Character. Name of attribute within \code{FRI} that will be stratified}

\item{features}{vector/list. Vector or list of vectors of features within \code{attribute}
to guide stratification.}

\item{raster}{spatRaster. Raster for polygon to raster conversion. If \code{map = TRUE}
raster must contain a layer named "strata".}

\item{map}{Logical. Default = \code{FALSE}. If \code{TRUE}, map values from FRIstrata 
with strata from \code{raster} to create secondary strata tier.}

\item{stack}{Logical. Default = \code{FALSE}. If \code{TRUE}, output raster will be
3 layers: \code{FRIstrata, rasterstrata, FRIrasterStrata}.}

\item{filename}{Character. Path to write stratified raster to disc.}

\item{overwrite}{Logical. Specify whether \code{filename} should be overwritten on disc.}

\item{plot}{Logical. Plots output spatRaster.}

\item{details}{Logical. If \code{FALSE} (default) output is spatRaster object of
stratified forest resources inventory attributes. If \code{TRUE} return a list
where \code{$outRaster} is the stratified forest resources inventory attributes,
\code{$lookUp} is the lookup table for the stratification, and \code{FRIpoly} is the
forest resources inventory poly with \code{attribute} and corresponding \code{class}}
}
\value{
A spatRaster object.
}
\description{
Stratify an input forest resources inventory polygon.
}
\seealso{
Other stratify functions: 
\code{\link{strat_breaks}()},
\code{\link{strat_kmeans}()},
\code{\link{strat_osb}()},
\code{\link{strat_pcomp}()},
\code{\link{strat_quantiles}()}
}
\concept{stratify functions}
