% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_kmeans.R
\name{strat_kmeans}
\alias{strat_kmeans}
\title{Stratify raster using k-means algorithm}
\usage{
strat_kmeans(
  raster,
  k,
  iter.max = 500,
  algorithm = "Lloyd",
  center = TRUE,
  scale = TRUE,
  plot = TRUE
)
}
\arguments{
\item{raster}{spatRaster. Raster to be used for simple random sampling}

\item{k}{Character. Number of k-mean centers to calculate.}

\item{iter.max}{Numeric. The maximum number of iterations allowed.}

\item{algorithm}{Character. \code{Lloyd} (default) or \code{MacQueen} algorithms}

\item{center}{Logical. Determines whether centering of data should be conducted prior to k-means analysis.}

\item{scale}{Logical. Determines whether scaling of data should be conducted prior to k-means analysis.}

\item{plot}{Logical. Plots output strata raster and visualized strata with boundary dividers.}
}
\value{
list where \code{kmeans} is all principal component analysis data and \code{raster} is the output stratification spatRaster
}
\description{
Stratify raster using k-means algorithm
}
\seealso{
Other stratify functions: 
\code{\link{strat_metrics}()},
\code{\link{strat_osb}()},
\code{\link{strat_pcomp}()}
}
\concept{stratify functions}
