% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_pcomp.R
\name{strat_pcomp}
\alias{strat_pcomp}
\title{Principal components stratification}
\usage{
strat_pcomp(
  mraster,
  nStrata,
  nStrata2 = NULL,
  center = TRUE,
  scale = TRUE,
  plot = FALSE,
  samp = 1,
  details = FALSE,
  filename = NULL,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{mraster}{spatRaster. ALS metrics raster.}

\item{nStrata}{Character. Number of desired strata.}

\item{nStrata2}{Numeric.  Number of secondary strata within \code{nStrata}.}

\item{center}{Logical. Value indicating whether the variables should be shifted to be zero centered.}

\item{scale}{Logical. Determines whether scaling and centering
of data should be conducted prior to analysis.}

\item{plot}{Logical. Plots output strata raster and visualized
strata with boundary dividers.}

\item{samp}{Numeric. For plotting - Determines proportion of cells
for strata visualization. Lower values reduce processing time.}

\item{details}{Logical. If \code{FALSE} (default) output is only
 stratification raster. If \code{TRUE} return a list
where \code{$details} is additional stratification information and
 \code{$raster} is the output stratification spatRaster.}

\item{filename}{Character. Path to write stratified raster to disc.}

\item{overwrite}{Logical. Specify whether \code{filename} should be overwritten on disc.}

\item{...}{Additional arguments to be passed to \code{\link[stats]{prcomp}}.}
}
\description{
Stratify metrics raster using principal components and quantile breaks
}
\seealso{
Other stratify functions: 
\code{\link{strat_breaks}()},
\code{\link{strat_kmeans}()},
\code{\link{strat_osb}()},
\code{\link{strat_quantiles}()}
}
\concept{stratify functions}
