% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataio.R
\name{summary.sgsObj}
\alias{summary.sgsObj}
\title{Prints summary of sgsObj data structure}
\usage{
\method{summary}{sgsObj}(object, ...)
}
\arguments{
\item{object}{An sgsObj to be summarized}

\item{...}{arguments passed to summary}
}
\value{
A printed summary of..
\itemize{
\item The number of individuals
\item The number of categories (analyses based on categories not yet implemented)
\item Number of loci
\item Number of alleles per locus
\item Number of gene copies per locus
\item The percentage of missing data per locus
}
}
\description{
Prints summary of sgsObj data structure
}
\examples{
## Simulate genetic data
Nind = 100 # Number of individuals
Nloci = 5 # Number of loci
Nallele = 10 # Number of alleles per loci

## Set up data frame and generate random spatial locations
dat <- data.frame(id = 0:(Nind - 1))
dat$x = runif(Nind, 0, 100)
dat$y = runif(Nind, 0, 100)

## Simulate Random genetic data and assign loci names
for (loci in 1:Nloci) {
 loci_name_a = paste("Loc", loci, "_A", sep = "")
 loci_name_b = paste("Loc", loci, "_B", sep = "")
 dat[loci_name_a] <- sample.int(Nallele, Nind, replace = TRUE)
 dat[loci_name_b] <- sample.int(Nallele, Nind, replace = TRUE)
}

## Convert to sgsObj
sgsObj = createSgsObj(sample_ids = dat$id,
                     genotype_data = dat[, 4:(Nloci*2 + 3)],
                     ploidy = 2,
                     x_coords = dat$x,
                     y_coords = dat$y)
summary(sgsObj)
}
\seealso{
\code{\link{createSgsObj}} for how to create a sgsObj
}

