% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_existing_strat.R
\name{take_samples}
\alias{take_samples}
\title{Take Samples Based on Strata}
\usage{
take_samples(existing, toSample, strata)
}
\arguments{
\item{existing}{A data frame containing existing data.}

\item{toSample}{A data frame specifying the desired sample sizes for each strata.}

\item{strata}{A string specifying the name of the variable used to define strata.}
}
\value{
A data frame containing a sample of the existing data based on the sample sizes
for the specified strata.
}
\description{
This function takes a data frame of existing data, a data frame of desired sample sizes
for each strata, and a strata variable name, and returns a sample of the existing data
based on the sample sizes for the specified strata.
}
\keyword{internal}
