% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tallySamples.R
\name{tallySamples}
\alias{tallySamples}
\title{Tally required number of samples for each raster strata}
\usage{
tallySamples(sraster, n, existing = NULL)
}
\arguments{
\item{sraster}{spatRaster. Stratification raster to be used for sampling.}

\item{n}{Numeric. Number of desired samples.}

\item{existing}{sf or data.frame.  Existing plot network.}
}
\value{
data.frame of strata and associates samples
}
\description{
Tally required number of samples for each raster strata
}
