% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-allocation.R
\name{allocating}
\alias{allocating}
\alias{allocation_prop}
\alias{allocation_optim}
\alias{allocation_manual}
\alias{allocation_equal}
\title{Allocating}
\usage{
allocation_prop(sraster, nSamp)

allocation_optim(sraster, mraster, nSamp)

allocation_manual(sraster, nSamp, weights)

allocation_equal(sraster, nSamp)
}
\arguments{
\item{sraster}{spatRaster. Stratification raster to be used for sampling.}

\item{nSamp}{Numeric. Number of desired samples.}

\item{mraster}{spatRaster. ALS metric raster. Required when \code{allocation = optim}.}

\item{weights}{Numeric. Only applicable when \code{allocation = "manual"}. Vector of weights where \code{sum(weights) == 1}. Vector length
must be equal to the number of unique strata where the first numeric value corresponds to stratum 1, second stratum 2 etc.}
}
\description{
Allocation algorithms
}
\concept{allocating}
\concept{allocation}
\keyword{internal}
