% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_allocation.R
\name{calculate_allocation}
\alias{calculate_allocation}
\title{Determine required samples in strata}
\usage{
calculate_allocation(
  sraster,
  nSamp,
  allocation = "prop",
  mraster = NULL,
  existing = NULL,
  force = FALSE
)
}
\arguments{
\item{sraster}{spatRaster. Stratification raster to be used for sampling.}

\item{nSamp}{Numeric. Number of desired samples.}

\item{allocation}{Character. Allocation algorithm to be used. Either \code{prop} (default) for proportional allocation
or \code{optim} for optimal allocation (equal sampling cost) or \code{equal} for equal number of samples (defined by \code{nSamp})
 for each strata.}

\item{mraster}{spatRaster. ALS metric raster. Required when \code{allocation = optim}.}

\item{existing}{sf 'POINT' or data.frame.  Existing plot network.}

\item{force}{Logical. \code{Default = FALSE} - force \code{nSamp} to be exactly the user defined value
in cases where \code{nSamp} and \code{sraster} strata count are not equally divisible. Additional samples often need to be allocated or removed
based on rounding differences resulting from proportional differences between \code{nSamp} and strata coverages in \code{sraster}.
In these instances samples are either added to strata with the lowest number of samples or are removed from strata with the highest number of samples.
Has no effect when \code{existing} is provided.}
}
\value{
Returns a data.frame of:
\itemize{
\item{strata} - Strata ID.
\item{total} - Total samples to be allocated based on under representation (positive) or over representation (negative)
to be removed at the users discretion.
\item{need} - Total required samples per strata. Rounded.
}
}
\description{
Determine required samples in strata
}
\details{
Determine how many samples to allocate within strata
}
\examples{
#--- Load strata raster and existing samples---#
r <- system.file("extdata", "kmeans.tif", package = "sgsR")
sr <- terra::rast(r)

e <- system.file("extdata", "existing.shp", package = "sgsR")
e <- sf::st_read(e)

#--- perform grid sampling ---#
calculate_allocation(
  sraster = sr,
  nSamp = 50
)

}
\references{
Gregoire, T.G., & Valentine, H.T. (2007). Sampling Strategies for Natural Resources and the Environment (1st ed.).
 Chapman and Hall/CRC. https://doi.org/10.1201/9780203498880
}
\seealso{
Other calculate functions: 
\code{\link{calculate_coobs}()},
\code{\link{calculate_distance}()},
\code{\link{calculate_pcomp}()},
\code{\link{calculate_pop}()},
\code{\link{calculate_representation}()},
\code{\link{calculate_sampsize}()}
}
\author{
Tristan R.H. Goodbody
}
\concept{calculate functions}
