% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_allocation.R
\name{calculate_allocation}
\alias{calculate_allocation}
\title{Determine required samples in strata}
\usage{
calculate_allocation(
  sraster,
  nSamp,
  allocation = "prop",
  mraster = NULL,
  existing = NULL,
  force = FALSE
)
}
\arguments{
\item{sraster}{spatRaster. Stratification raster to be used for sampling.}

\item{nSamp}{Numeric. Number of desired samples.}

\item{allocation}{Character. Allocation algorithm to be used. Either \code{prop} (default) for proportional allocation
or \code{optim} for optimal allocation (equal sampling cost) or \code{equal} for equal number of samples (defined by \code{nSamp})
 for each strata.}

\item{mraster}{spatRaster. ALS metric raster. Required when \code{allocation = optim}.}

\item{existing}{sf or data.frame.  Existing plot network.}

\item{force}{Logical. \code{Default = FALSE} - force \code{nSamp} to be exactly the user defined value
in cases where nSamp and \code{sraster} strata count are not equally divisible. Has no effect when \code{existing}
is provided.}
}
\value{
data.frame of:
\itemize{
\item{strata} - Strata ID.
\item{total} - Total samples to be allocated (positive) or removed (negative).
\item{need} - Total required samples per strata.
}
}
\description{
Determine required samples in strata
}
\details{
Determine how many samples to allocate within strata
}
\examples{
#--- Load strata raster and existing samples---#
r <- system.file("extdata", "kmeans.tif", package = "sgsR")
sr <- terra::rast(r)

e <- system.file("extdata", "existing.shp", package = "sgsR")
e <- sf::st_read(e)

#--- perform grid sampling ---#
calculate_allocation(
  sraster = sr,
  nSamp = 200
)

calculate_allocation(
  sraster = sr,
  nSamp = 200,
  force = TRUE
)

#--- extract strata from existing samples ---#
e.sr <- extract_strata(
  sraster = sr,
  existing = e
)

calculate_allocation(
  sraster = sr,
  nSamp = 200,
  existing = e.sr
)

#--- Load mraster for optimal allocation ---#
mr <- system.file("extdata", "wall_metrics.tif", package = "sgsR")
mr <- terra::rast(mr)

calculate_allocation(
  sraster = sr,
  nSamp = 200,
  existing = e.sr,
  allocation = "optim",
  mraster = mr$zq90,
  force = TRUE
)
}
\references{
Gregoire, T.G., & Valentine, H.T. (2007). Sampling Strategies for Natural Resources and the Environment (1st ed.).
 Chapman and Hall/CRC. https://doi.org/10.1201/9780203498880
}
\seealso{
Other calculate functions: 
\code{\link{calculate_coobs}()},
\code{\link{calculate_distance}()},
\code{\link{calculate_lhsPop}()},
\code{\link{calculate_pcomp}()},
\code{\link{calculate_sampsize}()}
}
\author{
Tristan R.H. Goodbody
}
\concept{calculate functions}
