% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_COOBS.R
\name{calculate_COOBS}
\alias{calculate_COOBS}
\title{COOBS algorithm sampling}
\usage{
calculate_COOBS(
  mraster = NULL,
  existing = NULL,
  cores = 1,
  threshold = 0.95,
  plot = FALSE,
  details = FALSE
)
}
\arguments{
\item{mraster}{spatRaster. ALS metrics raster. Requires at least 2 layers to calculate covariance matrix}

\item{existing}{sf. Samples resulting from sample_* functions.}

\item{cores}{Numeric. Number of CPU cores to use for parallel processing. \code{default = 1}}

\item{threshold}{Numeric. Proxy maximum pixel quantile to avoid outliers. \code{default = 0.95}}

\item{plot}{Logical. Plots output strata raster and visualized
strata with boundary dividers.}

\item{details}{Logical. If \code{FALSE} (default) output is only
 stratification raster. If \code{TRUE} return a list
where \code{$details} is additional stratification information and
 \code{$raster} is the output stratification spatRaster.}
}
\value{
output raster with COOBS and classified COOBS layers.
}
\description{
Perform the COunt of OBServations (COOBS) algorithm using existing site data
and raster metrics. This algorithm aids the user in determining where additional samples
could be located by comparing existing samples to each pixel and associated covariates.
The output COOBS raster could be used to constrain clhs sampling to areas that are underreprented.
}
\references{
Malone BP, Minansy B, Brungard C. 2019. Some methods to improve the utility of conditioned Latin hypercube sampling. PeerJ 7:e6451 DOI 10.7717/peerj.6451
}
\seealso{
Other calculate functions: 
\code{\link{calculate_distance}()},
\code{\link{calculate_pcomp}()},
\code{\link{calculate_reqSamples}()}
}
\concept{calculate functions}
