% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance.R
\name{calculate_distance}
\alias{calculate_distance}
\title{distance to access}
\usage{
calculate_distance(raster = NULL, access = NULL, plot = FALSE)
}
\arguments{
\item{raster}{spatRaster. Raster to be used to calculate pixel level distance to access layer.}

\item{access}{sf. Road access network - must be lines.}

\item{plot}{Logical. Plots output strata raster with samples.}
}
\value{
Input raster with \code{dist2access} layer appended.
}
\description{
Per pixel distance to nearest access vector. Intended to be used as a `cost` constraint
within the \code{\link{sample_clhs}} function
}
\seealso{
Other calculate functions: 
\code{\link{calculate_COOBS}()},
\code{\link{calculate_lhsPop}()},
\code{\link{calculate_pcomp}()},
\code{\link{calculate_reqSamples}()}
}
\concept{calculate functions}
