% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pcomp.R
\name{calculate_pcomp}
\alias{calculate_pcomp}
\title{Calculate and rasterize principal components from a metric raster}
\usage{
calculate_pcomp(
  mraster = NULL,
  nComp = NULL,
  scale = TRUE,
  plot = FALSE,
  details = FALSE,
  ...
)
}
\arguments{
\item{mraster}{spatRaster. ALS metrics raster.}

\item{nComp}{Numeric. Value indicating number of principal components to be rasterized.}

\item{scale}{Logical. Determines whether scaling and centering
of data should be conducted prior to analysis.}

\item{plot}{Logical. Plots output strata raster with samples.}

\item{details}{Logical. If \code{FALSE} (default) output is only
 stratification raster. If \code{TRUE} return a list
where \code{$details} is additional stratification information and
 \code{$raster} is the output stratification spatRaster.}
}
\value{
Output raster with specified number of principal components as layers.
}
\description{
Calculate and rasterize principal components from a metric raster
}
\seealso{
Other calculate functions: 
\code{\link{calculate_COOBS}()},
\code{\link{calculate_distance}()}
}
\concept{calculate functions}
