% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_representation.R
\name{calculate_representation}
\alias{calculate_representation}
\title{Compare representation of samples within sraster strata}
\usage{
calculate_representation(sraster, existing, plot = FALSE)
}
\arguments{
\item{sraster}{spatRaster. Stratification raster to be used for sampling.}

\item{existing}{sf or data.frame.  Existing plot network.}

\item{plot}{Logical. Plots existing (circles) and new (crosses) samples.}
}
\value{
Returns a data.frame of:
\itemize{
\item{strata} - \code{sraster} strata ID.
\item{srasterFreq} - Coverage frequency of \code{sraster} strata.
\item{sampleFreq} - Sampling frequency within \code{sraster} strata.
\item{diffFreq} - Difference between \code{srasterFreq} & \code{sampleFreq}. Positive values indicate over representation
\item{nSamp} - Number of samples within each strata in \code{existing}.
\item{need} - Total required samples to be representative of strata coverage. Rounded.
}
}
\description{
Compare representation of samples within sraster strata
}
\details{
Calculate how well sraster strata are represented in existing samples
}
\examples{
###--- generate example stratification ---###

#--- load ALS metrics from sgsR internal data ---#
r <- system.file("extdata", "wall_metrics.tif", package = "sgsR")

#--- read ALS metrics using the terra package ---#
mraster <- terra::rast(r)

#--- perform stratification ---#
sraster <- strat_kmeans(mraster = mrasters$zq90,
                        nStrata = 6,
                        plot = TRUE)


###--- create existing sample network ---###
set.seed(2022)

#--- load ALS metrics from sgsR internal data ---#
r <- system.file("extdata", "wall_metrics.tif", package = "sgsR")

#--- read ALS metrics using the terra package ---#
mraster <- terra::rast(r)

#--- simple random sampling ---#
existing <- sample_srs(raster = mraster$zq90,
                       nSamp = 45)
                       
#--- calculate representation ---#

calculate_representation(sraster = sraster, 
                         existing = existing, 
                         plot = TRUE)
                         
}
\seealso{
Other calculate functions: 
\code{\link{calculate_allocation}()},
\code{\link{calculate_coobs}()},
\code{\link{calculate_distance}()},
\code{\link{calculate_lhsPop}()},
\code{\link{calculate_pcomp}()},
\code{\link{calculate_sampsize}()}
}
\author{
Tristan R.H. Goodbody, Martin Queinnec
}
\concept{calculate functions}
