% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_reqSamples.R
\name{calculate_reqSamples}
\alias{calculate_reqSamples}
\title{Determine required samples in strata}
\usage{
calculate_reqSamples(sraster, nSamp, existing = NULL)
}
\arguments{
\item{sraster}{spatRaster. Stratification raster to be used for random sampling.}

\item{nSamp}{Numeric. Number of desired samples.}

\item{existing}{sf or data.frame.  Existing plot network.}
}
\value{
data.frame of strata and associates samples
}
\description{
Determine required samples in strata
}
\details{
Count the required number of samples for each raster strata
}
\examples{
\dontrun{
#--- Load strata raster and existing samples---#
r <- system.file("extdata","kmeans.tif", package = "sgsR")
sr <- terra::rast(r)

e <- system.file("extdata","existing.shp", package = "sgsR")
e <- sf::st_read(e)

#--- perform grid sampling ---#
calculate_reqSamples(sraster = sr, nSamp = 200)

calculate_reqSamples(sraster = sr, nSamp = 200, existing = e)
}
}
\seealso{
Other calculate functions: 
\code{\link{calculate_COOBS}()},
\code{\link{calculate_distance}()},
\code{\link{calculate_lhsPop}()},
\code{\link{calculate_pcomp}()}
}
\author{
Tristan R.H. Goodbody
}
\concept{calculate functions}
