% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_reqSamples.R
\name{calculate_reqSamples}
\alias{calculate_reqSamples}
\title{Determine required samples in strata}
\usage{
calculate_reqSamples(sraster, n, existing = NULL)
}
\arguments{
\item{sraster}{spatRaster. Stratification raster to be used for sampling.}

\item{n}{Numeric. Number of desired samples.}

\item{existing}{sf or data.frame.  Existing plot network.}
}
\value{
data.frame of strata and associates samples
}
\description{
Determine required samples in strata
}
\details{
Count the required number of samples for each raster strata
}
\seealso{
Other calculate functions: 
\code{\link{calculate_COOBS}()},
\code{\link{calculate_distance}()},
\code{\link{calculate_pcomp}()}
}
\concept{calculate functions}
