% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_existing.R
\name{extract_existing}
\alias{extract_existing}
\title{Extract raster cell values using existing samples}
\usage{
extract_existing(raster, existing, data.frame = FALSE)
}
\arguments{
\item{raster}{spatRaster. Multiband ALS metrics raster.}

\item{existing}{sf or data.frame.  Existing plot network to be included within sample_* functions}

\item{data.frame}{Logical. If true outputs as data.frame}
}
\value{
An sf or data.frame object with raster cell attributes
}
\description{
Extract raster cell values using existing samples
}
\seealso{
Other extract functions: 
\code{\link{extract_metrics}()}
}
\concept{extract functions}
