% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_existing.R
\name{extract_existing}
\alias{extract_existing}
\title{Extract raster strata to samples}
\usage{
extract_existing(sraster, existing, data.frame = FALSE)
}
\arguments{
\item{sraster}{spatRaster. Stratification raster to be used for sampling.}

\item{existing}{sf or data.frame.  Existing plot network.}

\item{data.frame}{Logical. If true outputs as data.frame}
}
\value{
An sf or data.frame object of samples with strata attributes
}
\description{
Extract stratification raster strata to existing samples
}
\seealso{
Other extract functions: 
\code{\link{extract_metrics}()}
}
\concept{extract functions}
