% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_metrics.R
\name{extract_metrics}
\alias{extract_metrics}
\title{Extract metric raster attributes to samples}
\usage{
extract_metrics(mraster, samples, data.frame = FALSE)
}
\arguments{
\item{mraster}{spatRaster. ALS metrics raster.}

\item{samples}{sf. Samples resulting from sample_* functions.}

\item{data.frame}{Logical. If true outputs as data.frame}
}
\value{
An sf or data.frame object of samples with associated raster cell attributes
}
\description{
Extract metric raster attributes to samples
}
\seealso{
Other extract functions: 
\code{\link{extract_existing}()}
}
\concept{extract functions}
