% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mask.R
\name{masking}
\alias{masking}
\alias{mask_access}
\title{Masking}
\usage{
mask_access(raster, access, buff_inner, buff_outer)
}
\arguments{
\item{raster}{SpatRaster. Raster to be masked.}

\item{access}{sf 'LINESTRING' or 'MULTILINESTRING'. Access network.}

\item{buff_inner}{Numeric. Inner buffer boundary specifying distance
from access where plots cannot be sampled.}

\item{buff_outer}{Numeric. Outer buffer boundary specifying distance
from access where plots can be sampled.}
}
\description{
Create covariate and sample matrices
}
\concept{masking}
\keyword{internal}
