% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot.sgsOut}
\alias{plot.sgsOut}
\title{Title}
\usage{
\method{plot}{sgsOut}(x, overlay = FALSE, color = "black",
  max_distance = NULL, ...)
}
\arguments{
\item{x}{}

\item{overlay}{}

\item{color}{}

\item{max_distance}{}

\item{...}{}
}
\value{
1
}
\description{
Title
}
\examples{
1
}

