% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-samplerules.R
\name{rules}
\alias{rules}
\alias{rule1}
\alias{rule2}
\title{Stratified sampling rules}
\usage{
rule1(n, s, i, strat_mask, add_strata, extraCols, mindist)

rule2(n, s, add_strata, nCount, strata_m, extraCols, mindist)
}
\arguments{
\item{mindist}{Numeric. Minimum allowable distance between selected
samples. \code{Default = NULL}.}
}
\description{
Stratified sampling rules

Stratified rule 1

Stratified rule 2
}
\concept{rules}
\keyword{internal}
