% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_aHELS.R
\name{sample_aHELS}
\alias{sample_aHELS}
\title{Adapted Hypercube Evaluation of a Legacy Sample (aHELS)}
\usage{
sample_aHELS(
  mraster,
  existing,
  nQuant = 10,
  nSamp = NULL,
  threshold = 0.9,
  plot = FALSE,
  filename = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{mraster}{spatRaster. ALS metrics raster.}

\item{existing}{sf. Samples resulting from sample_* functions.}

\item{nQuant}{Numeric. Number of quantiles to divide covariates and samples into.}

\item{nSamp}{Numeric. Maximum number of new samples to allocate. If provided, the algorithm
will default to allocating number of samples provided.}

\item{threshold}{Numeric. A sample quantile ratio threshold for establishing whether
additional samples should be added. \code{default = 0.9}. Values close to 1 can cause the algorithm to
continually loop and should be used sparingly.}

\item{plot}{Logial. Plots existing (circles) and new (crosses) samples on the first band of mraster.}

\item{filename}{Character. Path to write stratified raster to disc.}

\item{overwrite}{Logical. Specify whether \code{filename} should be overwritten on disc.}
}
\value{
Returns sf point object with existing samples and supplemental samples added by the aHELS algorithm.
}
\description{
Perform the adapted Hypercube Evaluation of a Legacy Sample (aHELS) algorithm using
existing site data and raster metrics. New samples are allocated based on quantile ratios between
the existing sampleand covariate dataset.
}
\note{
Special thanks to Brendan Malone for the original implementation of this algorithm.
}
\examples{
#--- Load raster and existing plots---#
r <- system.file("extdata","wall_metrics_small.tif", package = "sgsR")
mr <- terra::rast(r)

e <- system.file("extdata","existing.shp", package = "sgsR")
e <- sf::st_read(e)

sample_aHELS(mraster = mr[[1:3]], existing = e, plot = TRUE)

sample_aHELS(mraster = mr[[1:3]], existing = e, nQuant = 20, nSamp = 300, 
filename = tempfile(fileext = ".shp"))

}
\references{
Malone BP, Minansy B, Brungard C. 2019. Some methods to improve the utility of conditioned Latin hypercube sampling. PeerJ 7:e6451 DOI 10.7717/peerj.6451
}
\seealso{
Other sample functions: 
\code{\link{sample_balanced}()},
\code{\link{sample_clhs}()},
\code{\link{sample_grid}()},
\code{\link{sample_srs}()},
\code{\link{sample_strat}()}
}
\author{
Tristan R.H. Goodbody
}
\concept{sample functions}
