% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_ahels.R
\name{sample_ahels}
\alias{sample_ahels}
\title{Adapted Hypercube Evaluation of a Legacy Sample (ahels)}
\usage{
sample_ahels(
  mraster,
  existing,
  nQuant = 10,
  nSamp = NULL,
  threshold = 0.9,
  matrices = NULL,
  plot = FALSE,
  details = FALSE,
  filename = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{mraster}{spatRaster. ALS metrics raster.}

\item{existing}{sf.  Existing plot network.}

\item{nQuant}{Numeric. Number of quantiles to divide covariates and samples into. Quantiles that do not
cover at least 1 percent of the area of interest will be excluded and be returned as \code{NA}.}

\item{nSamp}{Numeric. Maximum number of new samples to allocate. If provided, the algorithm
will default to allocating number of samples provided.}

\item{threshold}{Numeric. A sample quantile ratio threshold for establishing whether
additional samples should be added. \code{default = 0.9}. Values close to 1 can cause the algorithm to
continually loop and should be used sparingly.}

\item{matrices}{List. Quantile and covariance matrices generated from \code{calculate_pop(mraster = mraster, nQuant = nQuant)}.
Both \code{mraster} & \code{nQuant} inputs must be the same to supply the covariance matrix. Supplying the matrix allows users
with very large rasters to pre-process the covariance matrix to avoid longer sampling processing times.}

\item{plot}{Logial. Plots existing (circles) and new (crosses) samples on the first band of mraster.}

\item{details}{Logical. If \code{FALSE} (default) output is only
 stratification raster. If \code{TRUE} return a list
where \code{$details} is additional stratification information and
 \code{$raster} is the output stratification spatRaster.
 @param ... Additional arguments to be passed to \code{\link[stats]{kmeans}} function.}

\item{filename}{Character. Path to write stratified raster to disc.}

\item{overwrite}{Logical. Specify whether \code{filename} should be overwritten on disc.}
}
\value{
Returns sf point object with existing samples and supplemental samples added by the ahels algorithm.
}
\description{
Perform the adapted Hypercube Evaluation of a Legacy Sample (ahels) algorithm using
existing site data and raster metrics. New samples are allocated based on quantile ratios between
the existing sample and covariate dataset.
}
\note{
Messages in the algorithm will state that samples have been added to under-represented quantiles. The number between
square brackets that follow represent the matrix row and column respectively that can be printed using \code{details = TRUE}.

In some cases, generally when a single metric is used as \code{mraster}, sampling ratios all be >= 1 before the 
\code{nSamp} number of samples are allocated. The algorithm will stop in this case.

Special thanks to Dr. Brendan Malone for the original implementation of this algorithm.
}
\examples{
#--- Load raster and existing plots---#
r <- system.file("extdata", "wall_metrics.tif", package = "sgsR")
mr <- terra::rast(r)

e <- system.file("extdata", "existing.shp", package = "sgsR")
e <- sf::st_read(e)

sample_ahels(
  mraster = mr,
  existing = e,
  plot = TRUE
)

#--- supply quantile and covariance matrices ---#
mat <- calculate_pop(mraster = mr)

sample_ahels(
  mraster = mr,
  existing = e,
  matrices = mat,
  nSamp = 300,
  filename = tempfile(fileext = ".shp")
)
}
\references{
Malone BP, Minansy B, Brungard C. 2019. Some methods to improve the utility of conditioned Latin hypercube sampling. PeerJ 7:e6451 DOI 10.7717/peerj.6451
}
\seealso{
Other sample functions: 
\code{\link{sample_balanced}()},
\code{\link{sample_clhs}()},
\code{\link{sample_srs}()},
\code{\link{sample_strat}()},
\code{\link{sample_systematic}()}
}
\author{
Tristan R.H. Goodbody
}
\concept{sample functions}
