% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_balanced.R
\name{sample_balanced}
\alias{sample_balanced}
\title{Balanced raster sampling using \code{\link{BalancedSampling}} and \code{\link{SamplingBigData}} methods}
\usage{
sample_balanced(
  mraster,
  n,
  algorithm = "lpm2",
  p = NULL,
  access = NULL,
  buff_inner = NULL,
  buff_outer = NULL,
  plot = FALSE,
  ...
)
}
\arguments{
\item{mraster}{spatRaster. ALS metrics raster.}

\item{n}{Numeric. Number of desired samples.}

\item{algorithm}{Character. One of \code{lpm2 lcube lcubestratified}}

\item{p}{Numeric. Inclusion probability for each candidate sample. 
Default is \code{n / N}}

\item{access}{sf. Road access network - must be lines.}

\item{buff_inner}{Numeric. Inner buffer boundary specifying distance
from access where plots cannot be sampled.}

\item{buff_outer}{Numeric. Outer buffer boundary specifying distance
from access where plots can be sampled.}

\item{plot}{Logical. Plots output strata raster and visualized
strata with boundary dividers.}

\item{...}{}
}
\value{
An sf object with \code{n} randomly sampled points.
}
\description{
Balanced raster sampling using \code{\link{BalancedSampling}} and \code{\link{SamplingBigData}} methods
}
\seealso{
Other sample functions: 
\code{\link{sample_aHELS}()},
\code{\link{sample_clhs}()},
\code{\link{sample_srs}()},
\code{\link{sample_strat}()}
}
\concept{sample functions}
