% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_grid.R
\name{sample_grid}
\alias{sample_grid}
\title{Grid sampling}
\usage{
sample_grid(
  raster,
  gridsize,
  access = NULL,
  buff_inner = NULL,
  buff_outer = NULL,
  plot = FALSE,
  filename = NULL,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{raster}{spatRaster. Raster used to define extent of fishnet grid}

\item{gridsize}{Numeric. Desired distance between samples}

\item{access}{sf. Road access network - must be lines.}

\item{buff_inner}{Numeric. Inner buffer boundary specifying distance
from access where plots cannot be sampled.}

\item{buff_outer}{Numeric. Outer buffer boundary specifying distance
from access where plots can be sampled.}

\item{plot}{Logical. Plots output strata raster with samples.}

\item{filename}{Character. Path to write output samples.}

\item{overwrite}{Logical. Choice to overwrite existing \code{filename} if it exists.}

\item{...}{Additional arguments for \link[sf]{st_make_grid}.}
}
\value{
An sf object with sampled points at intersections of fishnet grid.
}
\description{
Sample landscape using a fishnet grid pattern.
}
\seealso{
Other sample functions: 
\code{\link{sample_aHELS}()},
\code{\link{sample_balanced}()},
\code{\link{sample_clhs}()},
\code{\link{sample_srs}()},
\code{\link{sample_strat}()}
}
\concept{sample functions}
