% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_strat.R
\name{sample_strat}
\alias{sample_strat}
\title{Sampling based on a stratified raster}
\usage{
sample_strat(
  raster,
  ns,
  mindist,
  existing = NULL,
  include = FALSE,
  access = NULL,
  buff_inner = NULL,
  buff_outer = NULL,
  buff_extend = NULL,
  buff_max = NULL,
  wrow = 3,
  wcol = 3
)
}
\arguments{
\item{raster}{spatRaster. Raster to be used for simple random sampling}

\item{ns}{Numeric. Number of desired samples.}

\item{mindist}{Numeric. Minimum allowable distance between selected samples.}

\item{existing}{sf or data.frame.  Existing plot network}

\item{access}{sf. Road access network - must be lines.}

\item{buff_inner}{Numeric. Inner buffer boundary specifying distance from access where plots cannot be sampled.}

\item{buff_outer}{Numeric. Outer buffer boundary specifying distance from access where plots can be sampled.}

\item{buff_extend}{Numeric. Amount \code{buff_outer} will be increased by if no valid sample candidates are available.}

\item{buff_max}{Numeric. Maximum buffer distance resulting from increasing iterations of \code{buff_outer} + \code{buff_extend}.}

\item{wrow}{Numeric. Number of row in the focal window (default is 3).}

\item{wcol}{Numeric. Number of columns in the focal window (default is 3).}
}
\value{
list where \code{sampleDist} is a data.frame and \code{samples} is an sf object
}
\description{
Sampling based on a stratified raster
}
\examples{
sample_strat(raster = raster, ns = 200, mindist = 200)
sample_strat(raster = raster, ns = 200, mindist = 200, access = roads, buff_inner = 50, buff_outer = 200)
sample_strat(raster = raster, ns = 200, mindist = 200, access = roads, buff_inner = 50, buff_outer = 200, buff_extend = 100, buff_max = 800)

}
\seealso{
Other sample functions: 
\code{\link{sample_srs}()}
}
\concept{sample functions}
