% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_strat.R
\name{sample_strat}
\alias{sample_strat}
\title{Sampling based on a stratified raster.}
\usage{
sample_strat(
  sraster,
  n,
  mindist = 100,
  existing = NULL,
  include = FALSE,
  access = NULL,
  buff_inner = NULL,
  buff_outer = NULL,
  wrow = 3,
  wcol = 3,
  plot = FALSE,
  details = FALSE
)
}
\arguments{
\item{sraster}{spatRaster. Stratification raster to be used for sampling.}

\item{n}{Numeric. Number of desired samples.}

\item{mindist}{Numeric. Minimum allowable distance between selected samples. Default = 100.}

\item{existing}{sf or data.frame.  Existing plot network.}

\item{include}{Logical. If \code{TRUE} include existing plots in \code{n} total.}

\item{access}{sf. Road access network - must be lines.}

\item{buff_inner}{Numeric. Inner buffer boundary specifying distance from access where plots cannot be sampled.}

\item{buff_outer}{Numeric. Outer buffer boundary specifying distance from access where plots can be sampled.}

\item{wrow}{Numeric. Number of row in the focal window (default is 3).}

\item{wcol}{Numeric. Number of columns in the focal window (default is 3).}

\item{plot}{Logical. Plots output strata raster with samples.}

\item{details}{Logical. If \code{FALSE} (default) output is sf object of stratified samples. If \code{TRUE} return a list
where \code{$details} additional sampling information and \code{$raster} is an sf object of stratified samples.}
}
\value{
An sf object with \code{n} stratified samples.
}
\description{
Sampling based on a stratified raster.
}
\seealso{
Other sample functions: 
\code{\link{sample_balanced}()},
\code{\link{sample_srs}()}
}
\concept{sample functions}
