% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_systematic.R
\name{sample_systematic}
\alias{sample_systematic}
\title{Systematic sampling}
\usage{
sample_systematic(
  raster,
  cellsize,
  square = TRUE,
  centers = TRUE,
  access = NULL,
  buff_inner = NULL,
  buff_outer = NULL,
  plot = FALSE,
  filename = NULL,
  overwrite = FALSE,
  details = FALSE,
  ...
)
}
\arguments{
\item{raster}{spatRaster. Raster used to define extent of fishnet grid.}

\item{cellsize}{Numeric. Desired cellsize for tessellation.}

\item{square}{Logical. Tessellation shape. Default is regular square grid,
if \code{FALSE} hexagons are returned.}

\item{centers}{Logical. Sample location within tessellation. Default (\code{TRUE})
returns samples at centers. If \code{FALSE}, corners are returned.}

\item{access}{sf. Road access network - must be lines.}

\item{buff_inner}{Numeric. Inner buffer boundary specifying distance
from access where plots cannot be sampled.}

\item{buff_outer}{Numeric. Outer buffer boundary specifying distance
from access where plots can be sampled.}

\item{plot}{Logical. Plots output strata raster with samples.}

\item{filename}{Character. Path to write output samples.}

\item{overwrite}{Logical. Choice to overwrite existing \code{filename} if it exists.}

\item{details}{Logical. If \code{FALSE} (default) output is sf object of
systematic samples. If \code{TRUE} returns a list of sf objects where \code{tessellation}
is the tessellation grid for sampling, and \code{samples} are the systematic samples.}

\item{...}{Additional arguments for \code{\link[sf]{st_make_grid}}.}
}
\value{
An sf object with sampled points over a tessellation.
}
\description{
Systematic sampling within a square or hexagonal tessellation.
}
\examples{
#--- Load raster and access files ---#
r <- system.file("extdata", "kmeans.tif", package = "sgsR")
sr <- terra::rast(r)

a <- system.file("extdata", "roads.shp", package = "sgsR")
ac <- sf::st_read(a)

e <- system.file("extdata", "existing.shp", package = "sgsR")
e <- sf::st_read(e)

#--- perform grid sampling ---#
sample_systematic(
  raster = sr,
  cellsize = 1000,
  plot = TRUE
)

sample_systematic(
  raster = sr,
  cellsize = 1000,
  square = FALSE,
  centers = FALSE,
  plot = TRUE
)

sample_systematic(
  raster = sr,
  cellsize = 1000,
  access = ac,
  buff_inner = 50,
  buff_outer = 200
)

sample_systematic(
  raster = sr,
  cellsize = 1000,
  access = ac,
  buff_inner = 100,
  buff_outer = 400,
  filename = tempfile(fileext = ".shp"),
  plot = TRUE
)
}
\seealso{
Other sample functions: 
\code{\link{sample_ahels}()},
\code{\link{sample_balanced}()},
\code{\link{sample_clhs}()},
\code{\link{sample_srs}()},
\code{\link{sample_strat}()}
}
\author{
Tristan R.H. Goodbody
}
\concept{sample functions}
