% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genetic_functions.R
\name{sgs}
\alias{sgs}
\title{Run spatial genetic struture analysis}
\usage{
sgs(sgsObj, distance_intervals, nperm = 999)
}
\arguments{
\item{sgsObj}{A data structure created by \code{\link{createSgsObj}}}

\item{distance_intervals}{either a vector with upper limits of distance intervals, or a negative integer with number of distance intervals to generate with approximately equal pairwise comparisons}

\item{nperm}{Number of permutations of spatial locations among individuals to perform}
}
\value{
A list of the class 'sgsOut' that contains the following elements, which can be accessed with the $ operator (see examples):
\itemize{
\item sgsObj - the original sgsObj data structure used in the sgs analysis
\item di - summary of distance intervals used in analysis (see Details for more information)
\item fij_obs - observed Fij values in each distance interval for each locus, and averaged across loci
\item sp_obs - observed values of the Sp statistic in each distance interval for each locus, and averaged across loci
\item slope_obs - observed values of the slope of the regression of pairwise Fij on log distance for each distance interval and locus, and also averaged across loci
\item fij_perm_avg - average value of fij for each locus and distance interval after permutation of spatial locations among individuals
\item fij_perm_025 - lower 2.5\% quantile of fij for each locus and distance interval after permutation of spatial locations among individuals
\item fij_perm_975 - upper 97.5\% quantile of fij for each locus and distance interval after permutation of spatial locations among individuals

}
}
\description{
\code{sgs} is the primary function for analysis of spatial genetic structure in the sgsR package. This function takes the data structure generated by \code{\link{createSgsObj}} as its primary argument, runs an analysis of spatial genetic structure (see Details below), and outputs the results.
}
\section{Details}{


\emph{\strong{di}}

A summary of information about each distance interval. Includes:
\itemize{
\item Max distance - maximum pairwise distance within distance interval.
\item Average distance - average pairwise distance within distance interval.
\item Number of pairs - number of pairwise comparisons within distance interval.
\item \% participation - the proportion of all individuals represented at least once in each distance interval
\item CV participation - the coefficient of variation (CV) of the number of times each individual is represented in each distance interval
}

Values are checked against the rules of thumb presented by in the SPAGeDi manual and Vekemans and Hardy 2004 that the number of pairwise comparisons should be at least 100, % participation > 0.5, and CV participation <= 1 for each distance interval. Cautionary messages are printed if the choice of distance intervals violates one of these guidelines.
}
\examples{
## Simulate genetic data
Nind = 100 # Number of individuals
Nloci = 5 # Number of loci
Nallele = 10 # Number of alleles per loci

## Set up data frame and generate random spatial locations
dat <- data.frame(id = 0:(Nind - 1))
dat$x = runif(Nind, 0, 100)
dat$y = runif(Nind, 0, 100)

## Simulate Random genetic data and assign loci names
for (loci in 1:Nloci) {
 loci_name_a = paste("Loc", loci, "_A", sep = "")
 loci_name_b = paste("Loc", loci, "_B", sep = "")
 dat[loci_name_a] <- sample.int(Nallele, Nind, replace = TRUE)
 dat[loci_name_b] <- sample.int(Nallele, Nind, replace = TRUE)
}

## Convert to sgsObj
sgsObj = createSgsObj(sample_ids = dat$id,
                     genotype_data = dat[, 4:(Nloci*2 + 3)],
                     ploidy = 2,
                     x_coords = dat$x,
                     y_coords = dat$y)
summary(sgsObj)
}
\seealso{
\code{\link{createSgsObj}}, \code{\link{plot.sgsOut}}
}

