% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgsR-package.R
\docType{package}
\name{sgsR-package}
\alias{sgsR}
\alias{sgsR-package}
\title{sgsR: Structurally Guided Sampling}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Structurally guided sampling (SGS) approaches for airborne laser scanning (ALS; LIDAR). Primary functions provide means to generate data-driven stratifications & methods for allocating samples. Intermediate functions for calculating and extracting important information about input covariates and samples are also included. Processing outcomes are intended to help forest and environmental management practitioners better optimize field sample placement as well as assess and augment existing sample networks in the context of data distributions and conditions. ALS data is the primary intended use case, however any rasterized remote sensing data can be used, enabling data-driven stratifications and sampling approaches.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tgoodbody/sgsR}
  \item \url{https://tgoodbody.github.io/sgsR/}
  \item Report bugs at \url{https://github.com/tgoodbody/sgsR/issues}
}

}
\author{
\strong{Maintainer}: Tristan RH Goodbody \email{goodbody.t@gmail.com} (\href{https://orcid.org/0000-0002-6894-7925}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Nicholas C Coops \email{nicholas.coops@ubc.ca} (\href{https://orcid.org/0000-0002-0151-9037}{ORCID})
  \item Martin Queinnec \email{queinnec@mail.ubc.ca} (\href{https://orcid.org/0000-0002-2741-1032}{ORCID})
}

}
\keyword{internal}
