% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_breaks.R
\name{strat_breaks}
\alias{strat_breaks}
\title{Stratify metrics raster using user defined breaks}
\usage{
strat_breaks(
  mraster,
  metric = NULL,
  metric2 = NULL,
  breaks,
  breaks2 = NULL,
  plot = FALSE,
  details = FALSE
)
}
\arguments{
\item{mraster}{spatRaster. ALS metrics raster.}

\item{breaks}{Numeric. Vector of breakpoints for \code{metric}}

\item{breaks2}{Numeric. Vector of breakpoints for \code{metric2} (if provided)}

\item{plot}{Logical. Plots output strata raster and visualized
strata with boundary dividers.}

\item{details}{Logical. If \code{FALSE} (default) output is only
 stratification raster. If \code{TRUE} return a list
where \code{$details} is additional stratification information and
 \code{$raster} is the output stratification spatRaster.}
}
\value{
output stratification \code{spatRaster}, or a list when \code{details = TRUE}.
}
\description{
Stratify metrics raster using user defined breaks
}
\seealso{
Other stratify functions: 
\code{\link{strat_kmeans}()},
\code{\link{strat_osb}()},
\code{\link{strat_pcomp}()},
\code{\link{strat_quantiles}()}
}
\concept{stratify functions}
