% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_fri.R
\name{strat_fri}
\alias{strat_fri}
\title{Stratify Forest Resources Inventory (FRI)}
\usage{
strat_fri(
  fri,
  attribute,
  features,
  raster,
  filename = NULL,
  overwrite = FALSE,
  plot = FALSE,
  details = FALSE,
  ...
)
}
\arguments{
\item{fri}{sf. Forest resources inventory polygon coverage.}

\item{attribute}{Character. Name of attribute within \code{fri} that will be stratified}

\item{features}{vector/list. Vector or list of vectors of features within \code{attribute}
to guide stratification.}

\item{raster}{spatRaster. Raster for polygon to raster conversion. If \code{map = TRUE}
raster must contain a layer named "strata".}

\item{filename}{Character. Path to write stratified raster to disc.}

\item{overwrite}{Logical. Specify whether \code{filename} should be overwritten on disc.}

\item{plot}{Logical. Plots output spatRaster.}

\item{details}{Logical. If \code{FALSE} (default) output is spatRaster object of
stratified forest resources inventory attributes. If \code{TRUE} return a list
where \code{$outRaster} is the stratified forest resources inventory attributes,
\code{$lookUp} is the lookup table for the stratification, and \code{fripoly} is the
forest resources inventory poly with \code{attribute} and corresponding \code{strata}}

\item{...}{Additional arguments for writing files. See \code{\link[terra]{writeRaster}}.}
}
\value{
A spatRaster object.
}
\description{
Stratify an input forest resources inventory (FRI) polygon.
}
\seealso{
Other stratify functions: 
\code{\link{strat_breaks}()},
\code{\link{strat_kmeans}()},
\code{\link{strat_map}()},
\code{\link{strat_osb}()},
\code{\link{strat_pcomp}()},
\code{\link{strat_quantiles}()}
}
\concept{stratify functions}
