% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_kmeans.R
\name{strat_kmeans}
\alias{strat_kmeans}
\title{Stratify metrics raster using k-means algorithm}
\usage{
strat_kmeans(
  mraster,
  nstrata,
  iter.max = 500,
  algorithm = "Lloyd",
  scale = TRUE,
  plot = FALSE,
  details = FALSE
)
}
\arguments{
\item{mraster}{spatRaster. ALS metrics raster.}

\item{nstrata}{Character. Number of desired strata.}

\item{iter.max}{Numeric. The maximum number of iterations allowed.}

\item{algorithm}{Character. \code{Lloyd} (default) or 
\code{MacQueen} algorithms.}

\item{scale}{Logical. Determines whether scaling and centering
of data should be conducted prior to analysis.}

\item{plot}{Logical. Plots output strata raster and visualized
strata with boundary dividers.}

\item{details}{Logical. If \code{FALSE} (default) output is only
 stratification raster. If \code{TRUE} return a list
where \code{$details} is additional stratification information and
 \code{$raster} is the output stratification spatRaster.}
}
\value{
output stratification \code{spatRaster}, or a list when \code{details = TRUE}.
}
\description{
Stratify metrics raster using k-means algorithm
}
\seealso{
Other stratify functions: 
\code{\link{strat_breaks}()},
\code{\link{strat_osb}()},
\code{\link{strat_pcomp}()},
\code{\link{strat_quantiles}()}
}
\concept{stratify functions}
