% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_metrics.R
\name{strat_metrics}
\alias{strat_metrics}
\title{Stratify raster using metric quantiles}
\usage{
strat_metrics(
  raster,
  metric,
  metric2 = NULL,
  b,
  b2 = NULL,
  plot = TRUE,
  samp = 1
)
}
\arguments{
\item{raster}{spatRaster. Raster to be used for simple random sampling}

\item{metric}{Character. Name of primary metric to stratify}

\item{metric2}{Character. Name of secondary metric to stratify}

\item{b}{Numeric. Number of desired strata for metric}

\item{b2}{Numeric. Number of desired strata for metric2}

\item{plot}{Logical. Plots output strata raster and visualized strata with boundary dividers.}

\item{samp}{Numeric. Determines proportion of cells to plot for strata visualization. Lower values reduce processing time.}
}
\value{
list where \code{kmeans} is all principal component analysis data and \code{raster} is the output stratification spatRaster
}
\description{
Stratify raster using metric quantiles
}
\examples{
strat_metrics(raster = raster, metric = "p99", b = 4)
strat_metrics(raster = raster, metric = "p99", metric2 = "stdev", b = 4, b2 = 3)

}
\seealso{
Other stratify functions: 
\code{\link{strat_kmeans}()},
\code{\link{strat_osb}()},
\code{\link{strat_pcomp}()}
}
\concept{stratify functions}
