% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_osb.R
\name{strat_osb}
\alias{strat_osb}
\title{Determine optimum sample boundaries}
\usage{
strat_osb(
  mraster,
  metric = NULL,
  nStrata,
  nSamp,
  subset = 1,
  plot = FALSE,
  details = FALSE,
  filename = NULL,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{mraster}{spatRaster. ALS metrics raster.}

\item{metric}{Character. Name of metric to be used for stratification.}

\item{nStrata}{Numeric. Number of desired output strata.}

\item{nSamp}{Numeric. Number of desired samples - used within
OSB algorithm to help determine break points.}

\item{subset}{Numeric. Value between 0 and 1 (default)
denoting proportion of data to use to determine break points.}

\item{plot}{Logical. Plots output strata raster and visualized
strata with boundary dividers.}

\item{details}{Logical. If \code{FALSE} (default) output is only
 stratification raster. If \code{TRUE} return a list
where \code{$details} is additional stratification information and
 \code{$raster} is the output stratification spatRaster.
 @param ... Additional arguments to be passed to \code{\link[stats]{kmeans}} function.}

\item{filename}{Character. Path to write stratified raster to disc.}

\item{overwrite}{Logical. Specify whether \code{filename} should be overwritten on disc.}

\item{...}{Additional arguments for writing files. See \code{\link[terra]{writeRaster}}.}
}
\value{
Returns an output stratification \code{spatRaster} or a list when \code{details = TRUE}.

When a list is returned:
\enumerate{
\item \code{details} is a list output of the \code{\link[stratifyR]{strata.data}} function where
\code{OSB} are the optimum stratum boundaries and \code{nh} are the optimum sample sizes
for each stratum.
\item \code{osb} vector of optimum stratum boundaries.
\item \code{breaks} matrix associated metric and strata break values.
\item \code{raster} is a stratified \code{spatRaster} based on \code{OSB}.
}
}
\description{
Determine optimum sample boundaries algorithm of univariate populations
using the \code{\link[stratifyR]{strata.data}} algorithm.
}
\references{
Khan, E. A., Khan, M. G. M., & Ahsan, M. J. (2002). Optimum Stratification:
A Mathematical Programming Approach. Calcutta Statistical Association Bulletin,
52(1–4), 323–334. https://doi.org/10.1177/0008068320020518

Khan, M. G. M., Nand, N., & Ahmad, N. (2008). Determining the optimum strata
boundary points using dynamic programming. Survey methodology, 34(2), 205-214.

M.G.M. Khan, K.G. Reddy & D.K. Rao (2015) Designing stratified sampling in economic
and business surveys, Journal of Applied Statistics, 42:10, 2080-2099,
DOI: 10.1080/02664763.2015.1018674
}
\seealso{
Other stratify functions: 
\code{\link{strat_breaks}()},
\code{\link{strat_kmeans}()},
\code{\link{strat_pcomp}()},
\code{\link{strat_quantiles}()}
}
\concept{stratify functions}
