% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_osb.R
\name{strat_osb}
\alias{strat_osb}
\title{Optimum sample breaks stratification}
\usage{
strat_osb(
  mraster,
  metric = NULL,
  nStrata,
  nSamp,
  subset = 1,
  plot = FALSE,
  details = FALSE,
  filename = NULL,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{mraster}{spatRaster. ALS metrics raster.}

\item{metric}{Character. Name of metric to be used for stratification}

\item{nStrata}{Numeric. Number of desired output strata.}

\item{nSamp}{Numeric. Number of desired samples - used within
OSB algorithm to help determine break points.}

\item{subset}{- Numeric. Value between 0 and 1 (default)
denoting proportion of data to use to determine break points}

\item{plot}{Logical. Plots output strata raster and visualized
strata with boundary dividers.}

\item{details}{Logical. If \code{FALSE} (default) output is only
 stratification raster. If \code{TRUE} return a list
where \code{$details} is additional stratification information and
 \code{$raster} is the output stratification spatRaster.}

\item{filename}{Character. Path to write stratified raster to disc.}

\item{overwrite}{Logical. Specify whether \code{filename} should be overwritten on disc.}

\item{...}{Additional arguments for writing files. See \link[terra]{writeRaster}.}
}
\value{
output stratification \code{spatRaster}, or a list when \code{details = TRUE}.
}
\description{
Stratify metrics raster using optimum sample breaks algorithm
}
\seealso{
Other stratify functions: 
\code{\link{strat_breaks}()},
\code{\link{strat_kmeans}()},
\code{\link{strat_pcomp}()},
\code{\link{strat_quantiles}()}
}
\concept{stratify functions}
