% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_osb.R
\name{strat_osb}
\alias{strat_osb}
\title{Stratify metrics raster using optimum sample breaks algorithm}
\usage{
strat_osb(
  mraster,
  metric = NULL,
  nstrata,
  n,
  subset = 1,
  plot = FALSE,
  details = FALSE
)
}
\arguments{
\item{mraster}{spatRaster. ALS metrics raster.}

\item{metric}{Character. Name of metric to be used for stratification}

\item{nstrata}{Numeric. Number of desired output strata.}

\item{n}{Numeric. Number of desired samples - used within 
OSB algorithm to help determine break points.}

\item{subset}{- Numeric. Value between 0 and 1 (default) 
denoting proportion of data to use to determine break points}

\item{plot}{Logical. Plots output strata raster and visualized
strata with boundary dividers.}

\item{details}{Logical. If \code{FALSE} (default) output is only
 stratification raster. If \code{TRUE} return a list
where \code{$details} is additional stratification information and
 \code{$raster} is the output stratification spatRaster.}
}
\value{
output stratification \code{spatRaster}, or a list when \code{details = TRUE}.
}
\description{
Stratify metrics raster using optimum sample breaks algorithm
}
\seealso{
Other stratify functions: 
\code{\link{strat_breaks}()},
\code{\link{strat_kmeans}()},
\code{\link{strat_pcomp}()},
\code{\link{strat_quantiles}()}
}
\concept{stratify functions}
