% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_pcomp.R
\name{strat_pcomp}
\alias{strat_pcomp}
\title{Stratify raster using principal components and quantile breaks}
\usage{
strat_pcomp(raster, ncp, b1, b2 = NULL, scale = TRUE, plot = TRUE, samp = 1)
}
\arguments{
\item{raster}{spatRaster. Raster to be used for simple random sampling}

\item{ncp}{Character. Number of components to create.}

\item{b1}{Numeric. Number of desired strata for first principal component.}

\item{b2}{Numeric. Number of desired strata for second principal component.}

\item{scale}{Logical. Determines whether centering and scaling of data should be conducted prior to principal component analysis.}

\item{plot}{Logical. Plots output strata raster and visualized strata with boundary dividers.}

\item{samp}{Numeric. Determines proportion of cells to plot for strata visualization. Lower values reduce processing time.}
}
\value{
list where \code{pca} is all principal component analysis data and \code{raster} is the output stratification spatRaster
}
\description{
Stratify raster using principal components and quantile breaks
}
\examples{
strat_pcomp(raster = raster, ncp = 4, b1 = 4)
strat_pcomp(raster = raster, ncp = 4, b1 = 4, b2 = 3)

}
\seealso{
Other stratify functions: 
\code{\link{strat_kmeans}()},
\code{\link{strat_metrics}()},
\code{\link{strat_osb}()}
}
\concept{stratify functions}
