% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_pcomp.R
\name{strat_pcomp}
\alias{strat_pcomp}
\title{Principal components stratification}
\usage{
strat_pcomp(
  mraster,
  nStrata,
  nStrata2 = NULL,
  center = TRUE,
  scale = TRUE,
  plot = FALSE,
  samp = 1,
  details = FALSE,
  filename = NULL,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{mraster}{Spatraster. Covariate raster to stratify.}

\item{nStrata}{Numeric. Number of strata for primary principal component.}

\item{nStrata2}{Numeric. Number of strata for secondary principal component.}

\item{center}{Logical. Value indicating whether the variables should be shifted to be zero centered.}

\item{scale}{Logical. Value indicating whether the variables should be scaled to have unit variance}

\item{plot}{Logical. Plots output strata raster and visualized
strata with boundary dividers.}

\item{samp}{Numeric. Proportion of raster cells to plot between 0-1.}

\item{details}{Logical. If \code{FALSE} (default) output is only
 stratification raster. If \code{TRUE} return a list
where \code{$details} is additional stratification information and
 \code{$raster} is the output stratification spatRaster.
 @param ... Additional arguments to be passed to \code{\link[stats]{kmeans}} function.}

\item{filename}{Character. Path to write stratified raster to disc.}

\item{overwrite}{Logical. Specify whether \code{filename} should be overwritten on disc.}

\item{...}{Additional arguments to be passed to \code{\link[stats]{prcomp}}.}
}
\value{
Returns an output stratification \code{spatRaster} or a list when \code{details = TRUE}.

When a list is returned:
\enumerate{
\item \code{details} is a list output of the \code{\link[stats]{prcomp}} function
\item \code{raster} is a stratified \code{spatRaster} based on the PCA
\item \code{plot} is a \code{ggplot} scatter plot object where strata are colour coded
and strata boundaries are delineated
}
}
\description{
Stratify metrics raster using principal components and quantile breaks
}
\examples{
#--- Load raster and access files ---#
r <- system.file("extdata", "wall_metrics.tif", package = "sgsR")
mr <- terra::rast(r)

strat_pcomp(
  mraster = mr,
  nStrata = 5,
  plot = TRUE
)

strat_pcomp(
  mraster = mr,
  nStrata = 4,
  nStrata2 = 4,
  plot = TRUE,
  details = TRUE
)

strat_pcomp(
  mraster = mr,
  nStrata = 3,
  nStrata2 = 3,
  filename = tempfile(fileext = ".tif")
)
}
\seealso{
Other stratify functions: 
\code{\link{strat_breaks}()},
\code{\link{strat_kmeans}()},
\code{\link{strat_map}()},
\code{\link{strat_poly}()}
}
\author{
Tristan R.H. Goodbody
}
\concept{stratify functions}
