% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_quantiles.R
\name{strat_quantiles}
\alias{strat_quantiles}
\title{Quantiles stratification}
\usage{
strat_quantiles(
  mraster,
  metric = NULL,
  metric2 = NULL,
  nstrata,
  nstrata2 = NULL,
  plot = FALSE,
  samp = 1,
  details = FALSE,
  filename = NULL,
  ..deparseOpts
)
}
\arguments{
\item{mraster}{spatRaster. ALS metrics raster.}

\item{metric}{Character. Name of primary metric to stratify. If
\code{mraster} is has 1 layer it is taken as default.}

\item{metric2}{Character. Name of secondary metric to stratify.}

\item{nstrata}{Character. Number of desired strata.}

\item{nstrata2}{Numeric.  Number of secondary strata within \code{nstrata}.}

\item{plot}{Logical. Plots output strata raster and visualized
strata with boundary dividers.}

\item{samp}{Numeric. For plotting - Determines proportion of cells
for strata visualization. Lower values reduce processing time.}

\item{details}{Logical. If \code{FALSE} (default) output is only
 stratification raster. If \code{TRUE} return a list
where \code{$details} is additional stratification information and
 \code{$raster} is the output stratification spatRaster.}

\item{filename}{Character. Path to write stratified raster to disc.}
}
\value{
output stratification \code{spatRaster}, or a list when \code{details = TRUE}.
}
\description{
Stratify metric raster using metric quantiles.
}
\seealso{
Other stratify functions: 
\code{\link{strat_breaks}()},
\code{\link{strat_kmeans}()},
\code{\link{strat_osb}()},
\code{\link{strat_pcomp}()}
}
\concept{stratify functions}
