% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_quantiles.R
\name{strat_quantiles}
\alias{strat_quantiles}
\title{Quantiles stratification}
\usage{
strat_quantiles(
  mraster,
  metric = NULL,
  metric2 = NULL,
  nQuant,
  nQuant2 = NULL,
  plot = FALSE,
  details = FALSE,
  samp = 1,
  filename = NULL,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{mraster}{spatRaster. ALS metrics raster.}

\item{metric}{Numeric/Character. Index or name of primary covariate within \code{mraster} to stratify.}

\item{metric2}{Numeric/Character. Index or name of secondary covariate within \code{mraster} to stratify.}

\item{nQuant}{Numeric. Number of quantiles to stratify primary covariate.}

\item{nQuant2}{Numeric. Number of quantiles to stratify secondary covariate.}

\item{plot}{Logical. Plots output strata raster and visualized
strata with boundary dividers.}

\item{details}{Logical. If \code{FALSE} (default) output is only
 stratification raster. If \code{TRUE} return a list
where \code{$details} is additional stratification information and
 \code{$raster} is the output stratification spatRaster.
 @param ... Additional arguments to be passed to \code{\link[stats]{kmeans}} function.}

\item{samp}{Numeric. Determines proportion of cells to plot in scatterplot - see \code{values}
for strata visualization. Lower values reduce visualization time.}

\item{filename}{Character. Path to write stratified raster to disc.}

\item{overwrite}{Logical. Specify whether \code{filename} should be overwritten on disc.}

\item{...}{Additional arguments for writing files. See \code{\link[terra]{writeRaster}}.}
}
\value{
Returns an output stratification \code{spatRaster} or a list when \code{details = TRUE}.

When a list is returned:
\enumerate{
\item \code{details} is a list output of the \code{\link[stats]{prcomp}} function
\item \code{raster} is a stratified \code{spatRaster} based on quantiles
\item \code{plot} is a \code{ggplot} histogram / scatter plot object (depends on whether metric2 was supplied).
Histogram shows distribution and break points while scatter plot shows colour coded and strata boundaries.
}
}
\description{
Stratify metric raster using metric quantiles.
}
\examples{
#--- Load raster and existing plots---#
r <- system.file("extdata", "wall_metrics_small.tif", package = "sgsR")
mr <- terra::rast(r)

strat_quantiles(mraster = mr, 
                metric = 4, 
                nQuant = 10,
                plot = TRUE, 
                details = TRUE)

strat_quantiles(mraster = mr, 
                metric = "zsd",
                metric2 = "zq95", 
                nQuant = 3, 
                nQuant2 = 4)

strat_quantiles(mraster = mr, 
                metric = 1, 
                metric2 = "zsd",
                nQuant = 2, 
                nQuant2 = 2, 
                filename = tempfile(fileext = ".tif"))
                
}
\seealso{
Other stratify functions: 
\code{\link{strat_breaks}()},
\code{\link{strat_fri}()},
\code{\link{strat_kmeans}()},
\code{\link{strat_osb}()},
\code{\link{strat_pcomp}()}
}
\author{
Tristan R.H. Goodbody
}
\concept{stratify functions}
