% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-write.R
\name{write}
\alias{write}
\alias{write_samples}
\alias{write_samples_df}
\alias{write_raster}
\title{Write}
\usage{
write_samples(samples, filename = NULL, overwrite = FALSE)

write_samples_df(samples, filename = NULL, overwrite = FALSE)

write_raster(raster, filename = NULL, overwrite = FALSE)
}
\arguments{
\item{filename}{Character. Path to write output samples.}

\item{overwrite}{Logical. Choice to overwrite existing \code{filename} if it exists.}

\item{raster}{SpatRaster. Raster to be written to disc}
}
\value{
No explicit output but files are written to disc if specified.
}
\description{
Write

Write samples

Write table

Write raster
}
\concept{write}
\keyword{internal}
